refactor: extract detail sheet UI
