+++ app/src/main/java/com/kgs/calendar/ui/KgsCalendarApp.kt
@@ -2440,8 +2440,8 @@ internal val LocalMorphItemDay = compositionLocalOf<LocalDate?> { null }
 internal const val EnablePerItemMorph = false
 
 private val MorphContentFadeMs = MorphDurationMs * 3 / 10
-private val MorphDayEnter: EnterTransition = fadeIn(tween(MorphContentFadeMs, easing = MorphEasing))
-private val MorphDayExit: ExitTransition =
+internal val MorphDayEnter: EnterTransition = fadeIn(tween(MorphContentFadeMs, easing = MorphEasing))
+internal val MorphDayExit: ExitTransition =
     fadeOut(tween(MorphContentFadeMs, delayMillis = MorphDurationMs - MorphContentFadeMs, easing = MorphEasing))
 // The month cell (the morph's container source/target) now cross-fades gradually instead of
 // snapping, so the tapped day's lighter background doesn't vanish/appear abruptly.
@@ -2492,176 +2492,12 @@ internal val MorphItemBoundsTransform = androidx.compose.animation.BoundsTransfo
         tween(MorphDurationMs, easing = MorphEasing)
     }
 }
-private val MorphDayBoundsTransform = androidx.compose.animation.BoundsTransform { _, _ ->
+internal val MorphDayBoundsTransform = androidx.compose.animation.BoundsTransform { _, _ ->
     tween(MorphDurationMs, easing = MorphEasing)
 }
 
-private const val TaskDetailMorphDurationMs = 460
-private const val TaskDetailControlsRevealDelayMs = 220
-private val TaskDetailMorphBoundsTransform = androidx.compose.animation.BoundsTransform { _, _ ->
-    tween(TaskDetailMorphDurationMs, easing = MotionEmphasized)
-}
-
-private fun taskDetailMorphKey(task: TaskEntity): String =
-    "task-detail:${task.collectionHref}:${task.resourceHref}"
-
-private fun DetailSheet.transitionKey(): String = when (this) {
-    is DetailSheet.Event -> "event:${event.collectionHref}:${event.resourceHref}:${event.startsAtMillis}"
-    is DetailSheet.Task -> taskDetailMorphKey(task)
-}
-
-/**
- * Groups views that share a single persistent surface. The Day and 3-day views are the SAME
- * surface — the timeline — so switching between them is NOT a screen swap: it stays composed
- * and animates the day-column width (3 ? 1 or 1 ? 3), which makes the focused day widen/narrow
- * in place while the neighbouring days physically slide in/out. Every other view is a distinct
- * surface, reached by a cross-fade (or, for the month grid, a container-transform morph of the
- * tapped day). Driving [AnimatedContent] off the *group* instead of the raw view is what keeps
- * Day?3-day from cross-fading.
- */
-/**
- * Marks this composable as a shared element keyed by [key]. When the active view changes
- * (e.g. 3-day ? 1-day, or month-cell ? 1-day), the element with the same key in the
- * outgoing and incoming content morphs from one bounds to the other.
- *
- * The resize mode is [ScaleToBounds][SharedTransitionScope.ResizeMode.ScaleToBounds]
- * (Material's "container transform"): the content is laid out ONCE at its own target size
- * and then *scaled* as a single rigid block to fit the animating bounds. This is the key
- * to the Google-Calendar-style zoom — the whole day (grid lines, header, every card) grows
- * or shrinks together as one unit, instead of each child re-laying-out frame by frame
- * (RemeasureToBounds), which is what made contents drift to the bottom and snap at the end.
- * `FillBounds` lets width and height scale independently so the month?day case reads as a
- * zoom while the 3-day?1-day case reads as a pure column-width stretch.
- *
- * Pass [enabled] = false to register the key without participating in the morph, so that
- * only the *single* matching day in each view is a shared element (the others just cross-
- * fade with their scene). No-op entirely when not inside the morph container.
- */
-@Composable
-internal fun Modifier.morphBounds(
-    key: Any,
-    enabled: Boolean = true,
-    enter: EnterTransition = MorphDayEnter,
-    exit: ExitTransition = MorphDayExit,
-    overlayClip: SharedTransitionScope.OverlayClip = MorphRoundedClip,
-    overlayShape: Shape? = null,
-    boundsTransform: androidx.compose.animation.BoundsTransform = MorphDayBoundsTransform,
-    remeasureToBounds: Boolean = false,
-): Modifier {
-    val shared = LocalSharedTransitionScope.current ?: return this
-    val anim = LocalMorphAnimatedVisibilityScope.current ?: return this
-    return with(shared) {
-        // Always remember the state (stable slot) so toggling [enabled] across
-        // recompositions doesn't shift the composition's remember table.
-        val contentState = rememberSharedContentState(key)
-        if (!enabled) return@with this@morphBounds
-        this@morphBounds.sharedBounds(
-            sharedContentState = contentState,
-            animatedVisibilityScope = anim,
-            enter = enter,
-            exit = exit,
-            boundsTransform = boundsTransform,
-            resizeMode = if (remeasureToBounds) {
-                SharedTransitionScope.ResizeMode.RemeasureToBounds
-            } else {
-                SharedTransitionScope.ResizeMode.scaleToBounds(
-                    contentScale = ContentScale.FillBounds,
-                    alignment = Alignment.TopStart,
-                )
-            },
-            clipInOverlayDuringTransition = overlayShape?.let { OverlayClip(it) } ?: overlayClip,
-            placeholderSize = if (remeasureToBounds) {
-                SharedTransitionScope.PlaceholderSize.AnimatedSize
-            } else {
-                SharedTransitionScope.PlaceholderSize.ContentSize
-            },
-            // Keep the morphing element above the fading content of both scenes.
-            zIndexInOverlay = 1f,
-        )
-    }
-}
-
-@Composable
-internal fun Modifier.taskDetailMorph(key: String, cornerRadius: Dp): Modifier =
-    morphBounds(
-        key = key,
-        enter = EnterTransition.None,
-        exit = fadeOut(animationSpec = snap()),
-        overlayShape = RoundedCornerShape(cornerRadius),
-        boundsTransform = TaskDetailMorphBoundsTransform,
-        remeasureToBounds = true,
-    )
-
-@Composable
-private fun Modifier.taskDetailFadeOverlay(): Modifier {
-    val shared = LocalSharedTransitionScope.current ?: return this
-    return with(shared) {
-        this@taskDetailFadeOverlay.renderInSharedTransitionScopeOverlay(10f) { true }
-    }
-}
-
-@Composable
-internal fun rememberTaskDetailMorphCornerRadius(
-    enabled: Boolean,
-    ownRadius: Dp,
-    counterpartRadius: Dp,
-): Dp {
-    val morphScope = LocalMorphAnimatedVisibilityScope.current
-    if (!enabled || morphScope == null) return ownRadius
-    val radius by morphScope.transition.animateDp(
-        transitionSpec = { tween(TaskDetailMorphDurationMs, easing = MotionEmphasized) },
-        label = "taskDetailCornerRadius",
-    ) { state ->
-        if (state == EnterExitState.Visible) ownRadius else counterpartRadius
-    }
-    return radius
-}
-
-@Composable
-internal fun rememberTaskDetailMorphProgress(): Float {
-    val morphScope = LocalMorphAnimatedVisibilityScope.current ?: return 1f
-    val progress by morphScope.transition.animateFloat(
-        transitionSpec = { tween(TaskDetailMorphDurationMs, easing = MotionEmphasized) },
-        label = "taskDetailMorphProgress",
-    ) { state ->
-        if (state == EnterExitState.Visible) 1f else 0f
-    }
-    return progress
-}
-
-internal fun lerpFloat(start: Float, end: Float, fraction: Float): Float =
-    start + (end - start) * fraction.coerceIn(0f, 1f)
-
-internal fun phasedProgress(progress: Float, start: Float, end: Float): Float {
-    if (end <= start) return if (progress >= end) 1f else 0f
-    return ((progress - start) / (end - start)).coerceIn(0f, 1f)
-}
-
-private fun Modifier.taskDetailRevealLayout(
-    progress: Float,
-    offsetY: Dp,
-    alphaProgress: Float = progress,
-): Modifier = this
-    .layout { measurable, constraints ->
-        val placeable = measurable.measure(constraints)
-        val visibleHeight = (placeable.height * progress.coerceIn(0f, 1f)).roundToInt()
-        layout(placeable.width, visibleHeight) {
-            placeable.placeRelative(0, 0)
-        }
-    }
-    .graphicsLayer {
-        alpha = alphaProgress.coerceIn(0f, 1f)
-        translationY = -offsetY.toPx() * (1f - progress.coerceIn(0f, 1f))
-        clip = false
-    }
-
-// Rounds the morphing box's corners while it's lifted into the transition overlay, so the
-// growing/shrinking month-cell <-> day "box" has soft edges like the month cell it came from
-// (otherwise the solid background backing is a hard rectangle). Applied to both sides of the
-// morph so it rounds in both directions. Only active during the transition; the settled views
-// keep their own shapes.
 private val MorphCornerRadiusDp = 12.dp
-private val MorphRoundedClip: SharedTransitionScope.OverlayClip =
+internal val MorphRoundedClip: SharedTransitionScope.OverlayClip =
     object : SharedTransitionScope.OverlayClip {
         override fun getClipPath(
             sharedContentState: SharedTransitionScope.SharedContentState,
@@ -2840,1729 +2676,6 @@ private fun CreateMenuButton(icon: ImageVector, text: String, onClick: () -> Uni
             Text(text, color = onContainer, fontWeight = FontWeight.SemiBold, fontSize = 15.sp, maxLines = 1)
         }
     }
-}
-
-@Composable
-private fun ParentTaskPeek(
-    parentTask: TaskEntity,
-    taskColorMode: TaskColorMode,
-    onClick: () -> Unit,
-) {
-    val hierarchyLineColor = TaskHierarchyLine
-    val morphProgress = rememberTaskDetailMorphProgress()
-    val morphCornerRadius = rememberTaskDetailMorphCornerRadius(
-        enabled = true,
-        ownRadius = 14.dp,
-        counterpartRadius = 18.dp,
-    )
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(44.dp)
-            .zIndex(201f)
-            .clipToBounds()
-            .taskDetailMorph(taskDetailMorphKey(parentTask), morphCornerRadius),
-    ) {
-        Box(
-            modifier = Modifier
-                .matchParentSize()
-                .zIndex(0f)
-                .drawBehind {
-                    val x = TaskHierarchyStemInset.toPx()
-                    val overlap = TaskHierarchyLineOverlap.toPx()
-                    drawLine(
-                        color = hierarchyLineColor,
-                        start = Offset(x, size.height - 16.dp.toPx() * morphProgress),
-                        end = Offset(x, size.height + overlap),
-                        strokeWidth = 1.65.dp.toPx(),
-                        cap = StrokeCap.Round,
-                    )
-                },
-        )
-        TaskRow(
-            task = parentTask,
-            taskColorMode = taskColorMode,
-            onTaskStatusChanged = { _, _ -> onClick() },
-            prominent = true,
-            modifier = Modifier
-                .offset(y = (-15).dp),
-            outerHorizontalPadding = 0.dp,
-            outerVerticalPadding = 0.dp,
-            fixedCardHeight = 64.dp,
-            fadeTextAtBottom = true,
-            priorityMotionEnabled = false,
-            detailMorphFromHeader = true,
-            onClick = onClick,
-        )
-    }
-}
-
-@Composable
-private fun TaskDetailTopFadeOverlay(
-    visible: Boolean,
-    modifier: Modifier = Modifier,
-) {
-    val morphProgress = rememberTaskDetailMorphProgress()
-    val fadeAlpha = if (visible) {
-        phasedProgress(morphProgress, start = 0f, end = 0.34f)
-    } else {
-        0f
-    }
-    Box(
-        modifier = modifier
-            .fillMaxWidth()
-            .height(20.dp)
-            .zIndex(1000f)
-            .taskDetailFadeOverlay()
-            .background(
-                Brush.verticalGradient(
-                    0f to MaterialTheme.colorScheme.surface.copy(alpha = fadeAlpha),
-                    1f to Color.Transparent,
-                ),
-            ),
-    )
-}
-
-@Composable
-private fun TaskDetailHeaderCard(
-    task: TaskEntity,
-    taskColorMode: TaskColorMode,
-    hasSubtasks: Boolean,
-    collections: List<CollectionEntity>,
-    hiddenCollectionHrefs: Set<String>,
-    readOnlySource: Boolean,
-    burstKey: Int,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onEditTask: (TaskEntity) -> Unit,
-    onDuplicateTask: (TaskEntity) -> Unit,
-    onCopyTaskTo: (TaskEntity, String) -> Unit,
-    onDeleteTask: (String) -> Unit,
-    morphGeneration: Int,
-) {
-    val cardColor = Color(task.displayColor(taskColorMode))
-    val contentColor = if (cardColor.isDark()) Color.White else Color(0xFF1C1A18)
-    val morphCornerRadius = rememberTaskDetailMorphCornerRadius(
-        enabled = true,
-        ownRadius = 18.dp,
-        counterpartRadius = 14.dp,
-    )
-    val morphProgress = rememberTaskDetailMorphProgress()
-    val titleFontSize = lerpFloat(13f, 22f, morphProgress).sp
-    val titleLineHeight = lerpFloat(16f, 26f, morphProgress).sp
-    val subtitleFontSize = lerpFloat(11f, 12f, morphProgress).sp
-    val subtitleLineHeight = lerpFloat(14f, 14f, morphProgress).sp
-    val checkboxBoxSize = lerpFloat(30f, 42f, morphProgress).dp
-    val checkboxIconSize = lerpFloat(20f, 28f, morphProgress).dp
-    val checkboxContainerSize = lerpFloat(30f, 42f, morphProgress).dp
-    val rowStartPadding = lerpFloat(9f, 14f, morphProgress).dp
-    val rowEndPadding = lerpFloat(9f, 6f, morphProgress).dp
-    val rowVerticalPadding = lerpFloat(7f, 12f, morphProgress).dp
-    val actionSpaceWidth = lerpFloat(if (hasSubtasks) 30f else 0f, 96f, morphProgress).dp
-    val titleFontWeight = FontWeight(lerpFloat(500f, 600f, morphProgress).roundToInt())
-    val subtitleText = if (morphProgress < 0.52f) {
-        task.localizedTaskTimeLabel()
-    } else {
-        stringResource(if (task.isCompleted) R.string.completed_task else R.string.task)
-    }
-    var controlsVisible by remember(task.resourceHref) { mutableStateOf(morphGeneration == 0) }
-    LaunchedEffect(task.resourceHref, morphGeneration) {
-        if (morphGeneration == 0) {
-            controlsVisible = true
-        } else {
-            controlsVisible = false
-            delay(TaskDetailControlsRevealDelayMs.toLong())
-            controlsVisible = true
-        }
-    }
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .taskDetailMorph(taskDetailMorphKey(task), morphCornerRadius)
-            .zIndex(200f)
-            .graphicsLayer { clip = false }
-            .background(cardColor, RoundedCornerShape(morphCornerRadius)),
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .graphicsLayer { clip = false }
-                .padding(start = rowStartPadding, end = rowEndPadding, top = rowVerticalPadding, bottom = rowVerticalPadding),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(10.dp),
-        ) {
-            Box(Modifier.size(checkboxContainerSize), contentAlignment = Alignment.Center) {
-                CompletionBurst(burstKey, color = contentColor, modifier = Modifier.size(56.dp))
-                TaskStatusCheckbox(
-                    status = task.effectiveStatus(),
-                    tint = contentColor,
-                    onStatusChange = { onTaskStatusChanged(task.resourceHref, it) },
-                    boxSize = checkboxBoxSize,
-                    iconSize = checkboxIconSize,
-                )
-            }
-            Column(Modifier.weight(1f), verticalArrangement = Arrangement.spacedBy(2.dp)) {
-                SelectionContainer {
-                    Text(
-                        task.title,
-                        color = contentColor,
-                        fontSize = titleFontSize,
-                        lineHeight = titleLineHeight,
-                        fontWeight = titleFontWeight,
-                    )
-                }
-                AnimatedContent(
-                    targetState = subtitleText,
-                    transitionSpec = {
-                        fadeIn(tween(110, easing = MotionStandard)) togetherWith
-                            fadeOut(tween(110, easing = MotionStandardAccelerate))
-                    },
-                    label = "taskHeaderMorphSubtitle",
-                ) { text ->
-                    Text(
-                        text,
-                        color = contentColor.copy(alpha = 0.74f),
-                        fontSize = subtitleFontSize,
-                        lineHeight = subtitleLineHeight,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                }
-            }
-            Box(
-                modifier = Modifier.width(actionSpaceWidth),
-                contentAlignment = Alignment.CenterEnd,
-            ) {
-                if (hasSubtasks && morphProgress < 1f) {
-                    Icon(
-                        Icons.Default.KeyboardArrowDown,
-                        contentDescription = null,
-                        tint = contentColor,
-                        modifier = Modifier
-                            .align(Alignment.CenterEnd)
-                            .size(20.dp)
-                            .alpha(1f - phasedProgress(morphProgress, 0.18f, 0.68f)),
-                    )
-                }
-                TaskDetailHeaderActions(
-                    visible = controlsVisible,
-                    task = task,
-                    collections = collections,
-                    hiddenCollectionHrefs = hiddenCollectionHrefs,
-                    readOnlySource = readOnlySource,
-                    contentColor = contentColor,
-                    onEditTask = onEditTask,
-                    onDuplicateTask = onDuplicateTask,
-                    onCopyTaskTo = onCopyTaskTo,
-                    onDeleteTask = onDeleteTask,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun TaskDetailHeaderActions(
-    visible: Boolean,
-    task: TaskEntity,
-    collections: List<CollectionEntity>,
-    hiddenCollectionHrefs: Set<String>,
-    readOnlySource: Boolean,
-    contentColor: Color,
-    onEditTask: (TaskEntity) -> Unit,
-    onDuplicateTask: (TaskEntity) -> Unit,
-    onCopyTaskTo: (TaskEntity, String) -> Unit,
-    onDeleteTask: (String) -> Unit,
-) {
-    AnimatedVisibility(
-        visible = visible,
-        enter = fadeIn(tween(MotionShort, easing = MotionStandard)) +
-            scaleIn(initialScale = 0.9f, animationSpec = tween(MotionShort, easing = MotionEmphasized)),
-        exit = fadeOut(tween(90, easing = MotionStandardAccelerate)),
-    ) {
-        Row {
-            IconButton(onClick = { onEditTask(task) }) {
-                Icon(Icons.Default.Edit, contentDescription = stringResource(R.string.edit), tint = contentColor)
-            }
-            DetailOverflowMenu(
-                copyTargets = collections.filter { it.supportsTasks && it.href != task.collectionHref && !it.isReadOnlyForUi() },
-                hiddenCollectionHrefs = hiddenCollectionHrefs,
-                recurringScopes = false,
-                itemLabel = stringResource(R.string.task),
-                canDelete = !readOnlySource,
-                iconTint = contentColor,
-                onDuplicate = { onDuplicateTask(task) },
-                onCopyTo = { onCopyTaskTo(task, it) },
-                onDeleteThis = {},
-                onDeleteFollowing = {},
-                onDeleteAll = { onDeleteTask(task.resourceHref) },
-            )
-        }
-    }
-}
-
-@Composable
-private fun DetailSheetContent(
-    detail: DetailSheet,
-    collections: List<CollectionEntity>,
-    hiddenCollectionHrefs: Set<String>,
-    accounts: List<AccountEntity>,
-    problemResources: List<CalendarResourceEntity>,
-    taskColorMode: TaskColorMode,
-    eventFieldOrder: List<String>,
-    taskFieldOrder: List<String>,
-    autoLoadMapPreviews: Boolean,
-    accountEmails: List<String>,
-    allTasks: List<TaskEntity>,
-    taskMorphGeneration: Int,
-    taskMorphSourceHref: String?,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onTaskPriorityChanged: (String, Int) -> Unit,
-    onTaskProgressChanged: (String, Int) -> Unit,
-    onEventParticipationChanged: (String, String) -> Unit,
-    onEditEvent: (EventEntity) -> Unit,
-    onDuplicateEvent: (EventEntity) -> Unit,
-    onCopyEventTo: (EventEntity, String) -> Unit,
-    onDeleteEvent: (String, EventDeleteScope, Long) -> Unit,
-    onEditTask: (TaskEntity) -> Unit,
-    onDuplicateTask: (TaskEntity) -> Unit,
-    onCopyTaskTo: (TaskEntity, String) -> Unit,
-    onDeleteTask: (String) -> Unit,
-    onOpenSubtask: (TaskEntity, TaskEntity) -> Unit,
-    onOpenParentTask: (TaskEntity) -> Unit,
-    onAddSubtask: (TaskEntity) -> Unit,
-    onClose: () -> Unit,
-) {
-    val task = (detail as? DetailSheet.Task)?.task
-    val descendantTasks = remember(task?.resourceHref, allTasks) {
-        task?.let { allTasks.descendantsOf(it) }.orEmpty()
-    }
-    val descendantTasksByUid = remember(descendantTasks) {
-        descendantTasks.associateBy { it.collectionHref to it.uid }
-    }
-    val directSubtaskHasFollowing = remember(task?.uid, descendantTasks) {
-        val directChildren = task?.let { current ->
-            descendantTasks.filter { it.collectionHref == current.collectionHref && it.parentUid == current.uid }
-        }.orEmpty()
-        directChildren.mapIndexed { index, child -> child.uid to (index != directChildren.lastIndex) }.toMap()
-    }
-    fun directRootUidFor(descendant: TaskEntity): String {
-        val currentTask = task ?: return descendant.uid
-        var cursor = descendant
-        val seen = mutableSetOf(descendant.uid)
-        while (!cursor.parentUid.isNullOrBlank() && cursor.parentUid != currentTask.uid) {
-            val parent = descendantTasksByUid[cursor.collectionHref to cursor.parentUid] ?: break
-            if (!seen.add(parent.uid)) break
-            cursor = parent
-        }
-        return cursor.uid
-    }
-    val parentTask = remember(task?.parentUid, task?.collectionHref, allTasks) {
-        task?.parentUid?.takeIf { it.isNotBlank() }?.let { parentUid ->
-            allTasks.firstOrNull { it.collectionHref == task.collectionHref && it.uid == parentUid }
-        }
-    }
-    // Detail sheets always start with only the current task's direct children visible.
-    // Nested descendants can be expanded explicitly and ignore the app-wide list default.
-    val subtaskHierarchy = rememberTaskHierarchyPresentation(descendantTasks, expandedByDefault = false)
-    val sourceCollection = remember(detail, collections) {
-        collections.firstOrNull { collection ->
-            when (detail) {
-                is DetailSheet.Event -> collection.href == detail.event.collectionHref
-                is DetailSheet.Task -> collection.href == detail.task.collectionHref
-            }
-        }
-    }
-    val readOnlySource = sourceCollection?.isReadOnlyForUi() == true
-    val eventDetailCapabilities = sourceCollection?.eventEditorCapabilities() ?: EventEditorCapabilities.Full
-    var lastCompleted by remember(task?.resourceHref) { mutableStateOf(task?.isCompleted ?: false) }
-    var burstKey by remember(task?.resourceHref) { mutableStateOf(0) }
-    LaunchedEffect(task?.isCompleted) {
-        val current = task?.isCompleted ?: false
-        if (current && !lastCompleted) burstKey++
-        lastCompleted = current
-    }
-    val detailResourceHref = when (detail) {
-        is DetailSheet.Event -> detail.event.resourceHref
-        is DetailSheet.Task -> detail.task.resourceHref
-    }
-    val pendingMutation = pendingMutationFor(detailResourceHref)
-    val problemResource = remember(detailResourceHref, problemResources) {
-        problemResources.firstOrNull {
-            it.href == detailResourceHref && !it.syncError.isNullOrBlank()
-        }
-    }
-    val sheetScrollState = rememberScrollState()
-    Column(
-        Modifier
-            .fillMaxSize()
-            .verticalScroll(sheetScrollState)
-            .padding(horizontal = 22.dp, vertical = 18.dp),
-        verticalArrangement = Arrangement.spacedBy(14.dp),
-    ) {
-        when {
-            pendingMutation != null -> PendingSyncDetailBanner(pendingMutation)
-            problemResource != null -> SyncIssueDetailBanner(problemResource.syncError.orEmpty())
-        }
-        when (detail) {
-            is DetailSheet.Event -> {
-                Surface(
-                    shape = RoundedCornerShape(18.dp),
-                    color = WarmPeach.copy(alpha = 0.42f),
-                    modifier = Modifier.fillMaxWidth(),
-                ) {
-                    Row(
-                        modifier = Modifier.padding(start = 14.dp, end = 6.dp, top = 12.dp, bottom = 12.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(10.dp),
-                    ) {
-                        Box(Modifier.size(18.dp).clip(CircleShape).background(Color(detail.event.displayColor())))
-                        Column(Modifier.weight(1f), verticalArrangement = Arrangement.spacedBy(2.dp)) {
-                            Text(
-                                stringResource(R.string.event),
-                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                fontSize = 12.sp,
-                                lineHeight = 14.sp,
-                                fontWeight = FontWeight.SemiBold,
-                            )
-                            SelectionContainer {
-                                Text(
-                                    detail.event.title,
-                                    color = WarmInk,
-                                    fontSize = 22.sp,
-                                    lineHeight = 26.sp,
-                                    fontWeight = FontWeight.SemiBold,
-                                )
-                            }
-                        }
-                        val ev = detail.event
-                        val recurring = ev.isRecurring || !ev.recurrenceRule.isNullOrBlank()
-                        IconButton(onClick = { onEditEvent(ev) }) {
-                            Icon(Icons.Default.Edit, contentDescription = stringResource(R.string.edit), tint = WarmInk)
-                        }
-                        DetailOverflowMenu(
-                            copyTargets = collections.filter { it.supportsEvents && it.href != ev.collectionHref && !it.isReadOnlyForUi() },
-                            hiddenCollectionHrefs = hiddenCollectionHrefs,
-                            recurringScopes = recurring,
-                            itemLabel = stringResource(R.string.event),
-                            canDelete = !readOnlySource,
-                            onDuplicate = { onDuplicateEvent(ev) },
-                            onCopyTo = { onCopyEventTo(ev, it) },
-                            onDeleteThis = { onDeleteEvent(ev.resourceHref, EventDeleteScope.This, ev.startsAtMillis) },
-                            onDeleteFollowing = { onDeleteEvent(ev.resourceHref, EventDeleteScope.ThisAndFollowing, ev.startsAtMillis) },
-                            onDeleteAll = { onDeleteEvent(ev.resourceHref, EventDeleteScope.All, ev.startsAtMillis) },
-                        )
-                    }
-                }
-            }
-            is DetailSheet.Task -> {
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .graphicsLayer { clip = false },
-                ) {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .graphicsLayer { clip = false },
-                    ) {
-                        if (parentTask != null) {
-                        ParentTaskPeek(
-                            parentTask = parentTask,
-                            taskColorMode = taskColorMode,
-                            onClick = { onOpenParentTask(parentTask) },
-                        )
-                        TaskHierarchyConnectorStrip(
-                            stemInset = TaskHierarchyStemInset,
-                            modifier = Modifier.offset(y = (-3).dp),
-                            height = 3.dp,
-                            progress = rememberTaskDetailMorphProgress(),
-                        )
-                        }
-                        TaskDetailHeaderCard(
-                            task = detail.task,
-                            taskColorMode = taskColorMode,
-                            hasSubtasks = descendantTasks.any { it.parentUid == detail.task.uid },
-                            collections = collections,
-                            hiddenCollectionHrefs = hiddenCollectionHrefs,
-                            readOnlySource = readOnlySource,
-                            burstKey = burstKey,
-                            onTaskStatusChanged = onTaskStatusChanged,
-                            onEditTask = onEditTask,
-                            onDuplicateTask = onDuplicateTask,
-                            onCopyTaskTo = onCopyTaskTo,
-                            onDeleteTask = onDeleteTask,
-                            morphGeneration = taskMorphGeneration,
-                        )
-                        val detailMorphProgress = rememberTaskDetailMorphProgress()
-                        val hierarchyProgress = if (taskMorphGeneration > 0) {
-                            phasedProgress(detailMorphProgress, start = 0f, end = 0.82f)
-                        } else {
-                            1f
-                        }
-                        val hierarchyAlpha = if (taskMorphGeneration > 0) {
-                            phasedProgress(detailMorphProgress, start = 0f, end = 0.38f)
-                        } else {
-                            1f
-                        }
-                        Column(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .zIndex(150f)
-                                .taskDetailRevealLayout(
-                                    progress = hierarchyProgress,
-                                    offsetY = 18.dp,
-                                    alphaProgress = hierarchyAlpha,
-                                ),
-                        ) {
-                            if (subtaskHierarchy.entries.any { it.visible }) {
-                                TaskHierarchyConnectorStrip(
-                                    stemInset = TaskHierarchyStemInset,
-                                    progress = hierarchyProgress,
-                                )
-                            }
-                            subtaskHierarchy.entries.forEachIndexed { entryIndex, entry ->
-                                val virtualParentContinuation = if (directSubtaskHasFollowing[directRootUidFor(entry.task)] == true) {
-                                    setOf(0)
-                                } else {
-                                    emptySet()
-                                }
-                                val shiftedContinuationLevels = entry.continuationLevels.map { it + 1 }.toSet() + virtualParentContinuation
-                                val entryProgress = if (taskMorphGeneration > 0) {
-                                    phasedProgress(
-                                        detailMorphProgress,
-                                        start = 0.03f + entryIndex * 0.025f,
-                                        end = 0.42f + entryIndex * 0.025f,
-                                    )
-                                } else {
-                                    1f
-                                }
-                                AnimatedTaskHierarchyEntry(
-                                    entry = entry,
-                                    modifier = Modifier.graphicsLayer {
-                                        alpha = entryProgress
-                                        translationY = -12.dp.toPx() * (1f - entryProgress)
-                                        clip = false
-                                    },
-                                ) {
-                                    TaskRow(
-                                        task = entry.task,
-                                        taskColorMode = taskColorMode,
-                                        onTaskStatusChanged = onTaskStatusChanged,
-                                        prominent = true,
-                                        hierarchyDepth = entry.depth + 1,
-                                        hierarchyContinuationLevels = shiftedContinuationLevels,
-                                        hierarchyLastSibling = entry.lastSibling,
-                                        hasSubtasks = entry.hasChildren,
-                                        subtasksExpanded = entry.expanded,
-                                        onToggleSubtasks = { subtaskHierarchy.toggle(entry.task) },
-                                        outerHorizontalPadding = 0.dp,
-                                        outerVerticalPadding = 3.dp,
-                                        connectorStemInset = TaskHierarchyStemInset,
-                                        detailMorphKey = taskDetailMorphKey(entry.task),
-                                        detailMorphFromHeader = entry.task.resourceHref == taskMorphSourceHref,
-                                        onClick = { onOpenSubtask(detail.task, entry.task) },
-                                    )
-                                }
-                            }
-                            if (!readOnlySource) {
-                                TextButton(
-                                    onClick = { onAddSubtask(detail.task) },
-                                    modifier = Modifier
-                                        .align(Alignment.End)
-                                        .padding(top = 6.dp),
-                                ) {
-                                    Icon(Icons.Default.Add, contentDescription = null, modifier = Modifier.size(18.dp))
-                                    Spacer(Modifier.width(6.dp))
-                                    Text(stringResource(R.string.add_subtask))
-                                }
-                            }
-                        }
-                    }
-                    TaskDetailTopFadeOverlay(
-                        visible = parentTask != null,
-                        modifier = Modifier.align(Alignment.TopCenter),
-                    )
-                }
-            }
-        }
-        when (detail) {
-            is DetailSheet.Event -> {
-                EventMetadata(
-                    event = detail.event,
-                    accountEmails = accountEmails,
-                    fieldOrder = eventFieldOrder.filter { eventDetailCapabilities.allows(it) },
-                    readOnlySource = readOnlySource,
-                    autoLoadMapPreviews = autoLoadMapPreviews,
-                    onParticipationChanged = { partstat -> onEventParticipationChanged(detail.event.resourceHref, partstat) },
-                )
-                detail.sourceFootnoteText(collections, accounts)?.let { SourceFootnote(it) }
-                Spacer(Modifier.height(56.dp))
-            }
-            is DetailSheet.Task -> {
-                val detailMorphProgress = rememberTaskDetailMorphProgress()
-                val supportingProgress = if (taskMorphGeneration > 0) {
-                    phasedProgress(detailMorphProgress, start = 0.04f, end = 0.88f)
-                } else {
-                    1f
-                }
-                val supportingAlpha = if (taskMorphGeneration > 0) {
-                    phasedProgress(detailMorphProgress, start = 0.42f, end = 0.82f)
-                } else {
-                    1f
-                }
-                Column(
-                    modifier = Modifier.graphicsLayer {
-                        alpha = supportingAlpha
-                        translationY = 22.dp.toPx() * (1f - supportingProgress)
-                        clip = false
-                    },
-                    verticalArrangement = Arrangement.spacedBy(14.dp),
-                ) {
-                    TaskMetadata(
-                        task = detail.task,
-                        fieldOrder = taskFieldOrder,
-                        autoLoadMapPreviews = autoLoadMapPreviews,
-                        editable = !readOnlySource,
-                        onPriorityChanged = { onTaskPriorityChanged(detail.task.resourceHref, it) },
-                        onProgressChanged = { onTaskProgressChanged(detail.task.resourceHref, it) },
-                    )
-                    detail.sourceFootnoteText(collections, accounts)?.let { SourceFootnote(it) }
-                    Spacer(Modifier.height(56.dp))
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun SyncIssueDetailBanner(syncError: String) {
-    val container = if (MaterialTheme.colorScheme.background.isDark()) {
-        SyncPendingOrange.copy(alpha = 0.22f)
-    } else {
-        SyncPendingOrange.copy(alpha = 0.18f)
-    }
-    Surface(
-        shape = RoundedCornerShape(16.dp),
-        color = container,
-        border = BorderStroke(1.dp, SyncPendingOrange.copy(alpha = 0.58f)),
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 13.dp, vertical = 11.dp),
-            verticalAlignment = Alignment.Top,
-            horizontalArrangement = Arrangement.spacedBy(10.dp),
-        ) {
-            Icon(Icons.Default.WarningAmber, contentDescription = null, tint = SyncPendingOrange, modifier = Modifier.size(19.dp))
-            Column(verticalArrangement = Arrangement.spacedBy(3.dp)) {
-                Text(
-                    stringResource(R.string.sync_issue_title),
-                    color = WarmInk,
-                    fontSize = 13.sp,
-                    lineHeight = 16.sp,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    stringResource(R.string.sync_issue_item_body),
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontSize = 12.sp,
-                    lineHeight = 15.sp,
-                )
-                syncError.takeIf { it.isNotBlank() }?.let {
-                    Text(
-                        it,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.78f),
-                        fontSize = 11.sp,
-                        lineHeight = 14.sp,
-                        maxLines = 3,
-                        overflow = TextOverflow.Ellipsis,
-                    )
-                }
-            }
-        }
-    }
-}
-
-internal data class TaskHierarchyEntry(
-    val task: TaskEntity,
-    val depth: Int,
-    val hasChildren: Boolean,
-    val expanded: Boolean,
-    val visible: Boolean,
-    val continuationLevels: Set<Int>,
-    val lastSibling: Boolean,
-)
-
-internal data class TaskHierarchyPresentation(
-    val entries: List<TaskHierarchyEntry>,
-    val toggle: (TaskEntity) -> Unit,
-)
-
-internal val TaskHierarchyStemInset = 10.dp
-private val TaskHierarchyLineOverlap = 4.dp
-
-internal fun taskHierarchyLayerZ(depth: Int): Float = 100f - depth.coerceAtLeast(0)
-
-internal fun taskPriorityLayerZ(task: TaskEntity, animationsEnabled: Boolean): Float {
-    if (!animationsEnabled || task.isInactive()) return 0f
-    val intensity = taskPriorityIntensity(task.priority)
-    return if (intensity > 0f) 1_000f + intensity * 100f else 0f
-}
-
-@Composable
-internal fun rememberTaskHierarchyPresentation(
-    tasks: List<TaskEntity>,
-    expandedByDefault: Boolean,
-    defaultExpandedResourceHrefs: Set<String> = emptySet(),
-): TaskHierarchyPresentation {
-    val expansionOverrides = remember(tasks.map { it.resourceHref }.toSet(), expandedByDefault, defaultExpandedResourceHrefs) {
-        mutableStateMapOf<String, Boolean>()
-    }
-    val entries = remember(tasks, expandedByDefault, defaultExpandedResourceHrefs, expansionOverrides.toMap()) {
-        val distinctTasks = tasks.distinctBy { it.resourceHref }
-        val position = distinctTasks.withIndex().associate { it.value.resourceHref to it.index }
-        val byCollectionUid = distinctTasks.associateBy { it.collectionHref to it.uid }
-        val globallyUniqueByUid = distinctTasks.groupBy { it.uid }
-            .filterValues { it.size == 1 }
-            .mapValues { it.value.single() }
-
-        fun candidateParent(task: TaskEntity): TaskEntity? {
-            val parentUid = task.parentUid?.takeIf { it.isNotBlank() } ?: return null
-            return byCollectionUid[task.collectionHref to parentUid] ?: globallyUniqueByUid[parentUid]
-        }
-
-        val parentByResource = distinctTasks.associate { task ->
-            var parent = candidateParent(task)
-            val seen = mutableSetOf(task.resourceHref)
-            while (parent != null && seen.add(parent.resourceHref)) {
-                parent = candidateParent(parent)
-            }
-            task.resourceHref to if (parent == null) candidateParent(task) else null
-        }
-        val childrenByParent = distinctTasks
-            .mapNotNull { child -> parentByResource[child.resourceHref]?.resourceHref?.let { it to child } }
-            .groupBy({ it.first }, { it.second })
-            .mapValues { (_, children) -> children.sortedBy { position[it.resourceHref] } }
-        val roots = distinctTasks.filter { parentByResource[it.resourceHref] == null }
-        val out = mutableListOf<TaskHierarchyEntry>()
-        val emitted = mutableSetOf<String>()
-
-        fun append(
-            task: TaskEntity,
-            depth: Int,
-            visible: Boolean,
-            continuationLevels: Set<Int>,
-            lastSibling: Boolean,
-        ) {
-            if (!emitted.add(task.resourceHref)) return
-            val children = childrenByParent[task.resourceHref].orEmpty()
-            val expanded = expansionOverrides[task.resourceHref]
-                ?: if (task.resourceHref in defaultExpandedResourceHrefs) true else expandedByDefault
-            out += TaskHierarchyEntry(
-                task = task,
-                depth = depth,
-                hasChildren = children.isNotEmpty(),
-                expanded = expanded,
-                visible = visible,
-                continuationLevels = continuationLevels,
-                lastSibling = lastSibling,
-            )
-            children.forEachIndexed { index, child ->
-                val childLast = index == children.lastIndex
-                val childContinuationLevels = if (childLast) continuationLevels else continuationLevels + depth
-                append(
-                    task = child,
-                    depth = depth + 1,
-                    visible = visible && expanded,
-                    continuationLevels = childContinuationLevels,
-                    lastSibling = childLast,
-                )
-            }
-        }
-        roots.forEachIndexed { index, root ->
-            append(
-                task = root,
-                depth = 0,
-                visible = true,
-                continuationLevels = emptySet(),
-                lastSibling = index == roots.lastIndex,
-            )
-        }
-        distinctTasks.filterNot { it.resourceHref in emitted }.forEach {
-            append(it, 0, visible = true, continuationLevels = emptySet(), lastSibling = true)
-        }
-        out
-    }
-    return TaskHierarchyPresentation(
-        entries = entries,
-        toggle = { task ->
-            val current = expansionOverrides[task.resourceHref]
-                ?: if (task.resourceHref in defaultExpandedResourceHrefs) true else expandedByDefault
-            expansionOverrides[task.resourceHref] = !current
-        },
-    )
-}
-
-internal fun Modifier.taskHierarchyConnector(
-    depth: Int,
-    continuationLevels: Set<Int>,
-    lastSibling: Boolean,
-    branchEndInset: Dp = 0.dp,
-    stemInset: Dp = 9.dp,
-    progress: Float = 1f,
-    lineColor: Color,
-): Modifier {
-    if (depth <= 0 || progress <= 0.001f) return this
-    return drawBehind {
-        val indent = 18.dp.toPx()
-        val stemInsetPx = stemInset.toPx()
-        val resolvedProgress = progress.coerceIn(0f, 1f)
-        val overlapPx = TaskHierarchyLineOverlap.toPx() * resolvedProgress
-        val centerY = size.height / 2f
-        val color = lineColor.copy(alpha = resolvedProgress)
-        val stroke = 1.65.dp.toPx()
-        continuationLevels.forEach { level ->
-            val x = (level * indent) + stemInsetPx
-            drawLine(
-                color,
-                Offset(x, centerY + (-overlapPx - centerY) * resolvedProgress),
-                Offset(x, centerY + (size.height + overlapPx - centerY) * resolvedProgress),
-                strokeWidth = stroke,
-                cap = StrokeCap.Round,
-            )
-        }
-        val branchLevel = depth - 1
-        val branchX = (branchLevel * indent) + stemInsetPx
-        val branchBottom = if (lastSibling) centerY else size.height + overlapPx
-        drawLine(
-            color,
-            Offset(branchX, centerY + (-overlapPx - centerY) * resolvedProgress),
-            Offset(branchX, centerY + (branchBottom - centerY) * resolvedProgress),
-            strokeWidth = stroke,
-            cap = StrokeCap.Round,
-        )
-        val branchEndX = depth * indent + branchEndInset.toPx() + overlapPx
-        drawLine(
-            color,
-            Offset(branchX, centerY),
-            Offset(branchX + (branchEndX - branchX) * resolvedProgress, centerY),
-            strokeWidth = stroke,
-            cap = StrokeCap.Round,
-        )
-    }
-}
-
-internal fun Modifier.taskHierarchyParentTail(
-    depth: Int,
-    stemInset: Dp,
-    verticalPadding: Dp,
-    progress: Float,
-    lineColor: Color,
-): Modifier {
-    if (progress <= 0.001f) return this
-    return drawBehind {
-        val indent = 18.dp.toPx()
-        val x = depth * indent + stemInset.toPx()
-        val stroke = 1.65.dp.toPx()
-        val resolvedProgress = progress.coerceIn(0f, 1f)
-        val startY = (size.height - verticalPadding.toPx() - TaskHierarchyLineOverlap.toPx()).coerceIn(0f, size.height)
-        val endY = startY + (size.height - startY) * resolvedProgress
-        drawLine(
-            lineColor.copy(alpha = resolvedProgress),
-            Offset(x, startY),
-            Offset(x, endY),
-            strokeWidth = stroke,
-            cap = StrokeCap.Round,
-        )
-    }
-}
-
-@Composable
-private fun TaskHierarchyConnectorStrip(
-    stemInset: Dp,
-    modifier: Modifier = Modifier,
-    height: Dp = 3.dp,
-    progress: Float = 1f,
-) {
-    val lineColor = TaskHierarchyLine
-    Box(
-        modifier = modifier
-            .fillMaxWidth()
-            .height(height)
-            .drawBehind {
-                val x = stemInset.toPx()
-                val overlapPx = TaskHierarchyLineOverlap.toPx()
-                val stroke = 1.65.dp.toPx()
-                drawLine(
-                    lineColor,
-                    Offset(x, size.height + overlapPx - (size.height + overlapPx * 2f) * progress.coerceIn(0f, 1f)),
-                    Offset(x, size.height + overlapPx),
-                    strokeWidth = stroke,
-                    cap = StrokeCap.Round,
-                )
-            },
-    )
-}
-
-@Composable
-internal fun AnimatedTaskHierarchyEntry(
-    entry: TaskHierarchyEntry,
-    modifier: Modifier = Modifier,
-    content: @Composable () -> Unit,
-) {
-    val visibilityState = remember(entry.task.resourceHref) {
-        MutableTransitionState(entry.visible)
-    }
-    visibilityState.targetState = entry.visible
-    val exiting = visibilityState.currentState && !visibilityState.targetState
-    AnimatedVisibility(
-        visibleState = visibilityState,
-        enter = expandVertically(
-            expandFrom = Alignment.Top,
-            clip = false,
-            animationSpec = tween(MotionMedium, easing = MotionEmphasized),
-        ) + fadeIn(animationSpec = tween(MotionShort, easing = MotionStandard)),
-        exit = slideOutVertically(
-            animationSpec = tween(MotionMedium, easing = MotionEmphasized),
-        ) { -it / 6 } + shrinkVertically(
-            shrinkTowards = Alignment.Top,
-            clip = false,
-            animationSpec = tween(MotionMedium, easing = MotionEmphasized),
-        ),
-        modifier = modifier
-            .zIndex(
-                if (exiting) {
-                    -1f + entry.depth * 0.01f
-                } else {
-                    taskHierarchyLayerZ(entry.depth) + taskPriorityLayerZ(entry.task, PriorityAnimationsEnabled)
-                },
-            )
-            .graphicsLayer {
-                clip = false
-            },
-    ) {
-        val hierarchyExitProgress by transition.animateFloat(
-            transitionSpec = { tween(MotionMedium, easing = MotionEmphasized) },
-            label = "taskHierarchyExitProgress",
-        ) { state ->
-            if (state == EnterExitState.PostExit) 1f else 0f
-        }
-        CompositionLocalProvider(LocalTaskHierarchyExitProgress provides hierarchyExitProgress) {
-            content()
-        }
-    }
-}
-
-@Composable
-private fun PendingSyncDetailBanner(mutation: PendingMutationEntity) {
-    val container = if (MaterialTheme.colorScheme.background.isDark()) {
-        SyncPendingOrange.copy(alpha = 0.22f)
-    } else {
-        SyncPendingOrange.copy(alpha = 0.18f)
-    }
-    val body = when (mutation.action) {
-        MutationAction.Delete -> stringResource(R.string.pending_deleted)
-        MutationAction.Put -> stringResource(R.string.pending_changed)
-        else -> stringResource(R.string.pending_changed)
-    }
-    Surface(
-        shape = RoundedCornerShape(16.dp),
-        color = container,
-        border = BorderStroke(1.dp, SyncPendingOrange.copy(alpha = 0.58f)),
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = 13.dp, vertical = 11.dp),
-            verticalAlignment = Alignment.Top,
-            horizontalArrangement = Arrangement.spacedBy(10.dp),
-        ) {
-            Icon(Icons.Default.WarningAmber, contentDescription = null, tint = SyncPendingOrange, modifier = Modifier.size(19.dp))
-            Column(verticalArrangement = Arrangement.spacedBy(2.dp)) {
-                Text(
-                    stringResource(R.string.not_synced_yet),
-                    color = WarmInk,
-                    fontSize = 13.sp,
-                    lineHeight = 16.sp,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    body,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontSize = 12.sp,
-                    lineHeight = 15.sp,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun SourceFootnote(text: String) {
-    Text(
-        text = text,
-        color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.72f),
-        fontSize = 11.sp,
-        lineHeight = 14.sp,
-        modifier = Modifier.padding(start = 4.dp, end = 4.dp, top = 2.dp),
-    )
-}
-
-@Composable
-private fun DetailSheet.sourceFootnoteText(
-    collections: List<CollectionEntity>,
-    accounts: List<AccountEntity>,
-): String? {
-    val href = when (this) {
-        is DetailSheet.Event -> event.collectionHref
-        is DetailSheet.Task -> task.collectionHref
-    }
-    val collection = collections.firstOrNull { it.href == href } ?: return null
-    val account = accounts.firstOrNull { it.id == collection.accountId }
-    val source = when {
-        collection.href.isLocalCollectionHrefUi() -> stringResource(R.string.local_calendar)
-        collection.isAndroidProviderForUi() -> account?.displayName ?: stringResource(R.string.android_device_calendars)
-        collection.isReadOnlyForUi() -> account?.displayName ?: collection.displayName
-        else -> account?.displayName ?: account?.username ?: stringResource(R.string.source)
-    }
-    return stringResource(R.string.source_line, source, collection.displayName)
-}
-
-@Composable
-private fun DetailOverflowMenu(
-    copyTargets: List<CollectionEntity>,
-    hiddenCollectionHrefs: Set<String> = emptySet(),
-    recurringScopes: Boolean,
-    itemLabel: String,
-    canDelete: Boolean = true,
-    iconTint: Color = WarmInk,
-    onDuplicate: () -> Unit,
-    onCopyTo: (String) -> Unit,
-    onDeleteThis: () -> Unit,
-    onDeleteFollowing: () -> Unit,
-    onDeleteAll: () -> Unit,
-) {
-    var expanded by remember { mutableStateOf(false) }
-    var copyExpanded by remember { mutableStateOf(false) }
-    var deleteDialogOpen by remember { mutableStateOf(false) }
-    Box {
-        IconButton(onClick = { expanded = true }) {
-            Icon(Icons.Default.MoreVert, contentDescription = stringResource(R.string.more_options), tint = iconTint)
-        }
-        DropdownMenu(
-            expanded = expanded,
-            onDismissRequest = { expanded = false },
-            shape = RoundedCornerShape(18.dp),
-            containerColor = MaterialTheme.colorScheme.surface,
-            shadowElevation = 8.dp,
-        ) {
-            DropdownMenuItem(
-                text = { Text(stringResource(R.string.duplicate)) },
-                onClick = {
-                    expanded = false
-                    onDuplicate()
-                },
-            )
-            DropdownMenuItem(
-                text = { Text(stringResource(R.string.copy_to)) },
-                enabled = copyTargets.isNotEmpty(),
-                onClick = {
-                    expanded = false
-                    copyExpanded = true
-                },
-            )
-            if (canDelete) {
-                DropdownMenuItem(
-                    text = { Text(stringResource(R.string.delete), color = MaterialTheme.colorScheme.error) },
-                    onClick = {
-                        expanded = false
-                        deleteDialogOpen = true
-                    },
-                )
-            }
-        }
-        if (copyExpanded) {
-            CopyTargetDialog(
-                copyTargets = copyTargets,
-                hiddenCollectionHrefs = hiddenCollectionHrefs,
-                onDismiss = { copyExpanded = false },
-                onCopyTo = { href ->
-                    copyExpanded = false
-                    onCopyTo(href)
-                },
-            )
-        }
-        if (deleteDialogOpen) {
-            DeleteConfirmationDialog(
-                itemLabel = itemLabel,
-                recurringScopes = recurringScopes,
-                onDismiss = { deleteDialogOpen = false },
-                onDeleteThis = {
-                    deleteDialogOpen = false
-                    onDeleteThis()
-                },
-                onDeleteFollowing = {
-                    deleteDialogOpen = false
-                    onDeleteFollowing()
-                },
-                onDeleteAll = {
-                    deleteDialogOpen = false
-                    onDeleteAll()
-                },
-            )
-        }
-    }
-}
-
-@Composable
-private fun CopyTargetDialog(
-    copyTargets: List<CollectionEntity>,
-    hiddenCollectionHrefs: Set<String> = emptySet(),
-    onDismiss: () -> Unit,
-    onCopyTo: (String) -> Unit,
-) {
-    AlertDialog(
-        modifier = Modifier.padding(horizontal = 20.dp),
-        onDismissRequest = onDismiss,
-        shape = RoundedCornerShape(24.dp),
-        title = {
-            Text(stringResource(R.string.copy_to), color = WarmInk, fontWeight = FontWeight.SemiBold)
-        },
-        text = {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .heightIn(max = 420.dp)
-                    .verticalScroll(rememberScrollState()),
-                verticalArrangement = Arrangement.spacedBy(4.dp),
-            ) {
-                copyTargets.forEach { collection ->
-                    val hidden = collection.href in hiddenCollectionHrefs
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clip(RoundedCornerShape(14.dp))
-                            .then(if (hidden) Modifier.dashedBorder(SyncPendingOrange, 14.dp) else Modifier)
-                            .clickable { onCopyTo(collection.href) }
-                            .padding(horizontal = 12.dp, vertical = 12.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(14.dp),
-                    ) {
-                        Box(Modifier.size(12.dp).clip(CircleShape).background(Color(collection.color)))
-                        Text(
-                            collection.displayName,
-                            color = WarmInk,
-                            fontSize = 16.sp,
-                            lineHeight = 19.sp,
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis,
-                        )
-                    }
-                }
-            }
-        },
-        confirmButton = {},
-        dismissButton = {
-            TextButton(onClick = onDismiss) { Text(appString(R.string.cancel)) }
-        },
-    )
-}
-
-/** Which slice of a recurring event a delete action targets. */
-private enum class EventDeleteScope { This, ThisAndFollowing, All }
-
-@Composable
-private fun RecurringSaveScopeDialog(
-    itemLabel: String,
-    onDismiss: () -> Unit,
-    onSaveThis: () -> Unit,
-    onSaveFollowing: () -> Unit,
-    onSaveAll: () -> Unit,
-) {
-    AlertDialog(
-        modifier = Modifier.padding(horizontal = 20.dp),
-        onDismissRequest = onDismiss,
-        shape = RoundedCornerShape(24.dp),
-        containerColor = scopeDialogBackground(),
-        titleContentColor = WarmInk,
-        textContentColor = WarmInk,
-        title = { Text(appString(R.string.edit_repeat_question), fontWeight = FontWeight.SemiBold) },
-        text = {
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                Text(
-                    appString(R.string.edit_series_body, itemLabel),
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontSize = 14.sp,
-                    lineHeight = 18.sp,
-                )
-                Spacer(Modifier.height(2.dp))
-                DeleteScopeButton(appString(R.string.only_this_repeat), onClick = onSaveThis)
-                DeleteScopeButton(appString(R.string.this_and_following), onClick = onSaveFollowing)
-                DeleteScopeButton(appString(R.string.all_repeats), onClick = onSaveAll)
-            }
-        },
-        confirmButton = {},
-        dismissButton = { TextButton(onClick = onDismiss) { Text(appString(R.string.cancel)) } },
-    )
-}
-
-@Composable
-private fun DeleteConfirmationDialog(
-    itemLabel: String,
-    recurringScopes: Boolean,
-    onDismiss: () -> Unit,
-    onDeleteThis: () -> Unit,
-    onDeleteFollowing: () -> Unit,
-    onDeleteAll: () -> Unit,
-) {
-    if (recurringScopes) {
-        // Step 1: pick a scope. Step 2: confirm that choice before anything is deleted.
-        var pending by remember { mutableStateOf<EventDeleteScope?>(null) }
-        val pendingScope = pending
-        if (pendingScope == null) {
-            // A recurring event offers three choices, laid out as a clean vertical stack of
-            // full-width buttons instead of cramming them into the dialog's button row.
-            AlertDialog(
-                modifier = Modifier.padding(horizontal = 20.dp),
-                onDismissRequest = onDismiss,
-                shape = RoundedCornerShape(24.dp),
-                containerColor = scopeDialogBackground(),
-                titleContentColor = WarmInk,
-                textContentColor = WarmInk,
-                title = { Text(appString(R.string.delete_repeat_question), fontWeight = FontWeight.SemiBold) },
-                text = {
-                    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                        Text(
-                            appString(R.string.event_belongs_to_series_delete),
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                            fontSize = 14.sp,
-                            lineHeight = 18.sp,
-                        )
-                        Spacer(Modifier.height(2.dp))
-                        DeleteScopeButton(appString(R.string.only_this_event), onClick = { pending = EventDeleteScope.This })
-                        DeleteScopeButton(appString(R.string.this_and_following_events), onClick = { pending = EventDeleteScope.ThisAndFollowing })
-                        DeleteScopeButton(appString(R.string.all_events_in_series), destructive = true, onClick = { pending = EventDeleteScope.All })
-                    }
-                },
-                confirmButton = {},
-                dismissButton = { TextButton(onClick = onDismiss) { Text(appString(R.string.cancel)) } },
-            )
-        } else {
-            AlertDialog(
-                modifier = Modifier.padding(horizontal = 20.dp),
-                onDismissRequest = { pending = null },
-                shape = RoundedCornerShape(24.dp),
-                containerColor = scopeDialogBackground(),
-                titleContentColor = WarmInk,
-                textContentColor = WarmInk,
-                title = { Text(appString(R.string.truly_delete), fontWeight = FontWeight.SemiBold) },
-                text = {
-                    Text(
-                        when (pendingScope) {
-                            EventDeleteScope.This -> appString(R.string.only_this_deleted)
-                            EventDeleteScope.ThisAndFollowing -> appString(R.string.this_and_following_deleted)
-                            EventDeleteScope.All -> appString(R.string.all_series_deleted)
-                        },
-                    )
-                },
-                confirmButton = {
-                    TextButton(
-                        onClick = {
-                            when (pendingScope) {
-                                EventDeleteScope.This -> onDeleteThis()
-                                EventDeleteScope.ThisAndFollowing -> onDeleteFollowing()
-                                EventDeleteScope.All -> onDeleteAll()
-                            }
-                        },
-                    ) { Text(appString(R.string.delete), color = MaterialTheme.colorScheme.error) }
-                },
-                dismissButton = { TextButton(onClick = { pending = null }) { Text(appString(R.string.cancel)) } },
-            )
-        }
-    } else {
-        AlertDialog(
-            modifier = Modifier.padding(horizontal = 20.dp),
-            onDismissRequest = onDismiss,
-            shape = RoundedCornerShape(24.dp),
-            containerColor = scopeDialogBackground(),
-            titleContentColor = WarmInk,
-            textContentColor = WarmInk,
-            title = { Text(appString(R.string.delete_confirm_question, itemLabel), fontWeight = FontWeight.SemiBold) },
-            text = { Text(appString(R.string.delete_irreversible)) },
-            confirmButton = {
-                TextButton(onClick = onDeleteAll) {
-                    Text(appString(R.string.delete), color = MaterialTheme.colorScheme.error)
-                }
-            },
-            dismissButton = { TextButton(onClick = onDismiss) { Text(appString(R.string.cancel)) } },
-        )
-    }
-}
-
-@Composable
-private fun DeleteScopeButton(text: String, destructive: Boolean = false, onClick: () -> Unit) {
-    val container = if (destructive) {
-        MaterialTheme.colorScheme.errorContainer
-    } else {
-        MaterialTheme.colorScheme.surface
-    }
-    val outline = if (destructive) {
-        MaterialTheme.colorScheme.error.copy(alpha = 0.46f)
-    } else {
-        MaterialTheme.colorScheme.primary.copy(alpha = 0.38f)
-    }
-    Surface(
-        onClick = onClick,
-        shape = RoundedCornerShape(16.dp),
-        color = container,
-        border = BorderStroke(1.dp, outline),
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        Text(
-            text,
-            modifier = Modifier.padding(horizontal = 16.dp, vertical = 14.dp),
-            color = if (destructive) MaterialTheme.colorScheme.onErrorContainer else MaterialTheme.colorScheme.onSurface,
-            fontSize = 15.sp,
-            lineHeight = 18.sp,
-            fontWeight = FontWeight.Medium,
-        )
-    }
-}
-
-@Composable
-private fun scopeDialogBackground(): Color =
-    popupSurfaceColor()
-
-@Composable
-private fun EventMetadata(
-    event: EventEntity,
-    accountEmails: List<String>,
-    fieldOrder: List<String>,
-    readOnlySource: Boolean,
-    autoLoadMapPreviews: Boolean,
-    onParticipationChanged: (String) -> Unit,
-) {
-    val organizer = remember(event.organizerJson) { event.organizerJson.toCalendarParticipant() }
-    val attendees = remember(event.attendeesJson) { event.attendeesJson.toCalendarParticipants() }
-    val accountEmailSet = remember(accountEmails) { accountEmails.map { it.lowercase(Locale.ROOT) }.toSet() }
-    val ownAttendee = attendees.firstOrNull { it.email.lowercase(Locale.ROOT) in accountEmailSet }
-    Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-        fieldOrder.forEach { field ->
-            when (field) {
-                "time" -> DetailLine(stringResource(R.string.time), event.localizedFullTimeLabel())
-                "location" -> event.location?.takeIf { it.isNotBlank() }?.let { LocationDetailLine(it, event.locationMapVerified, autoLoadMapPreviews) }
-                "participants" -> if (organizer != null || attendees.isNotEmpty()) {
-                    SharedEventDetail(
-                        organizer = organizer,
-                        attendees = attendees,
-                        ownAttendee = ownAttendee,
-                        onParticipationChanged = onParticipationChanged,
-                    )
-                }
-                "notes" -> event.description?.takeIf { it.isNotBlank() }?.let { DetailLine(stringResource(R.string.notes), it) }
-                "categories" -> event.categories?.takeIf { it.isNotBlank() }?.let { CategoryDetailLine(stringResource(R.string.categories), it) }
-                "status" -> if (!readOnlySource) {
-                    DetailLine(stringResource(R.string.event_status), event.statusLabelText())
-                    DetailLine(stringResource(R.string.sharing_visibility), event.classLabelText())
-                    DetailLine(stringResource(R.string.availability), event.transparencyLabelText())
-                }
-                "recurrence" -> event.recurrenceRule?.takeIf { it.isNotBlank() }?.let { DetailLine(stringResource(R.string.recurrence), it.toLocalizedRecurrenceLabel()) }
-                "reminders" -> event.remindersCsv.localizedReminderSummary()?.let { DetailLine(stringResource(R.string.reminders), it) }
-            }
-        }
-        event.syncError?.takeIf { it.isNotBlank() }?.let { DetailLine(stringResource(R.string.sync), it) }
-    }
-}
-
-@Composable
-private fun SharedEventDetail(
-    organizer: CalendarParticipant?,
-    attendees: List<CalendarParticipant>,
-    ownAttendee: CalendarParticipant?,
-    onParticipationChanged: (String) -> Unit,
-) {
-    Surface(
-        shape = RoundedCornerShape(14.dp),
-        color = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.38f),
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        Column(
-            modifier = Modifier.padding(horizontal = 14.dp, vertical = 11.dp),
-            verticalArrangement = Arrangement.spacedBy(10.dp),
-        ) {
-            Text(stringResource(R.string.participants), color = MaterialTheme.colorScheme.onSurfaceVariant, fontSize = 12.sp, lineHeight = 14.sp, fontWeight = FontWeight.SemiBold)
-            organizer?.let {
-                Text(stringResource(R.string.organized_by, it.displayName), color = WarmInk, fontSize = 14.sp, lineHeight = 17.sp, fontWeight = FontWeight.Medium)
-            }
-            attendees.forEach { attendee ->
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                ) {
-                    ParticipantCircle(
-                        label = attendee.initials(),
-                        color = attendee.avatarColor(),
-                        contentColor = Color.White,
-                        borderColor = attendee.displayParticipationColor(),
-                        modifier = Modifier.size(34.dp),
-                    )
-                    Column(Modifier.weight(1f)) {
-                        SelectionContainer {
-                            Text(attendee.displayName, color = WarmInk, fontSize = 14.sp, lineHeight = 17.sp, maxLines = 1, overflow = TextOverflow.Ellipsis)
-                        }
-                        Text(attendee.email, color = MaterialTheme.colorScheme.onSurfaceVariant, fontSize = 12.sp, lineHeight = 14.sp, maxLines = 1, overflow = TextOverflow.Ellipsis)
-                    }
-                    Text(attendee.localizedDisplayParticipationLabel(), color = attendee.displayParticipationColor(), fontSize = 12.sp, lineHeight = 14.sp, fontWeight = FontWeight.SemiBold)
-                }
-            }
-            ownAttendee?.let {
-                HorizontalDivider(color = WarmLine.copy(alpha = 0.72f))
-                Text(stringResource(R.string.my_participation), color = WarmInk, fontSize = 14.sp, lineHeight = 17.sp, fontWeight = FontWeight.SemiBold)
-                FadedHorizontalScrollRow(contentPadding = PaddingValues(horizontal = 0.dp), horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                    listOf("ACCEPTED" to stringResource(R.string.accept), "TENTATIVE" to stringResource(R.string.tentative), "DECLINED" to stringResource(R.string.decline)).forEach { (value, label) ->
-                        FilterChip(
-                            selected = it.partstat.equals(value, ignoreCase = true),
-                            onClick = { onParticipationChanged(value) },
-                            label = { Text(label) },
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun TaskMetadata(
-    task: TaskEntity,
-    fieldOrder: List<String>,
-    autoLoadMapPreviews: Boolean,
-    editable: Boolean,
-    onPriorityChanged: (Int) -> Unit,
-    onProgressChanged: (Int) -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-        fieldOrder.forEach { field ->
-            when (field) {
-                "status" -> DetailLine(stringResource(R.string.status), task.localizedStatusLabel())
-                "time" -> DetailLine(stringResource(R.string.time), task.localizedTaskTimeLabel())
-                "location" -> task.location?.takeIf { it.isNotBlank() }?.let { LocationDetailLine(it, task.locationMapVerified, autoLoadMapPreviews) }
-                "notes" -> task.notes?.takeIf { it.isNotBlank() }?.let { DetailLine(stringResource(R.string.note), it) }
-                "url" -> task.url?.takeIf { it.isNotBlank() }?.let { DetailLine(stringResource(R.string.url), it) }
-                "priority" -> task.priority?.let {
-                    if (editable) EditablePriorityLine(it, onPriorityChanged) else ReadOnlyPriorityLine(it)
-                }
-                "progress" -> {
-                    val progress = task.displayProgress()
-                    if (editable) EditableProgressLine(progress, onProgressChanged) else ReadOnlyProgressLine(progress)
-                }
-                "tags" -> task.categories?.takeIf { it.isNotBlank() }?.let { CategoryDetailLine(stringResource(R.string.categories), it) }
-                "recurrence" -> task.recurrenceRule?.takeIf { it.isNotBlank() }?.let { DetailLine(stringResource(R.string.recurrence), it.toLocalizedRecurrenceLabel()) }
-                "reminders" -> task.remindersCsv.localizedReminderSummary()?.let { DetailLine(stringResource(R.string.reminders), it) }
-            }
-        }
-        task.completedAtMillis?.let { DetailLine(stringResource(R.string.completed_on), it.toDate().toString()) }
-        task.syncError?.takeIf { it.isNotBlank() }?.let { DetailLine(stringResource(R.string.sync), it) }
-    }
-}
-
-@Composable
-private fun EditablePriorityLine(priority: Int, onPriorityChanged: (Int) -> Unit) {
-    var draftPriority by remember(priority) { mutableStateOf(priority.coerceIn(1, 9)) }
-    DetailMetricLine(label = stringResource(R.string.priority), value = draftPriority.coerceIn(1, 9).toString()) {
-        PrioritySlider(
-            priority = draftPriority,
-            onPriorityChange = { draftPriority = it },
-            onPriorityChangeFinished = onPriorityChanged,
-        )
-    }
-}
-
-@Composable
-private fun EditableProgressLine(progress: Int, onProgressChanged: (Int) -> Unit) {
-    var draftProgress by remember(progress) { mutableStateOf(progress.coerceIn(0, 100)) }
-    DetailMetricLine(label = stringResource(R.string.progress), value = "${draftProgress.coerceIn(0, 100)}%") {
-        ProgressSlider(
-            progress = draftProgress,
-            onProgressChange = { draftProgress = it },
-            onProgressChangeFinished = onProgressChanged,
-        )
-    }
-}
-
-@Composable
-private fun DetailLine(label: String, value: String) {
-    val displayValue = remember(value) { value.cleanCalendarDisplayText() }
-    Surface(
-        shape = RoundedCornerShape(14.dp),
-        color = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.38f),
-        modifier = Modifier
-            .fillMaxWidth()
-            .animateContentSize(animationSpec = tween(MotionMedium, easing = MotionStandard)),
-    ) {
-        Column(
-            modifier = Modifier.padding(horizontal = 14.dp, vertical = 11.dp),
-            verticalArrangement = Arrangement.spacedBy(4.dp),
-        ) {
-            Text(label, color = MaterialTheme.colorScheme.onSurfaceVariant, fontSize = 12.sp, lineHeight = 14.sp, fontWeight = FontWeight.SemiBold)
-            LinkableSelectableText(displayValue)
-        }
-    }
-}
-
-@Composable
-private fun ReadOnlyProgressLine(progress: Int) {
-    DetailMetricLine(label = stringResource(R.string.progress), value = "${progress.coerceIn(0, 100)}%") {
-        ReadOnlyProgressBar(progress = progress, modifier = Modifier.fillMaxWidth())
-    }
-}
-
-@Composable
-private fun ReadOnlyPriorityLine(priority: Int) {
-    DetailMetricLine(label = stringResource(R.string.priority), value = priority.coerceIn(1, 9).toString()) {
-        ReadOnlyPriorityBar(priority = priority, modifier = Modifier.fillMaxWidth())
-    }
-}
-
-@Composable
-private fun DetailMetricLine(
-    label: String,
-    value: String,
-    content: @Composable ColumnScope.() -> Unit,
-) {
-    Surface(
-        shape = RoundedCornerShape(14.dp),
-        color = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.38f),
-        modifier = Modifier
-            .fillMaxWidth()
-            .animateContentSize(animationSpec = tween(MotionMedium, easing = MotionStandard)),
-    ) {
-        Column(
-            modifier = Modifier.padding(horizontal = 14.dp, vertical = 11.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            Row(horizontalArrangement = Arrangement.SpaceBetween, verticalAlignment = Alignment.CenterVertically, modifier = Modifier.fillMaxWidth()) {
-                Text(label, color = MaterialTheme.colorScheme.onSurfaceVariant, fontSize = 12.sp, lineHeight = 14.sp, fontWeight = FontWeight.SemiBold)
-                Text(value, color = WarmInk, fontSize = 14.sp, lineHeight = 17.sp, fontWeight = FontWeight.SemiBold)
-            }
-            content()
-        }
-    }
-}
-
-@Composable
-private fun ReadOnlyProgressBar(progress: Int, modifier: Modifier = Modifier) {
-    val normalized = progress.coerceIn(0, 100)
-    val animated by animateFloatAsState(
-        targetValue = normalized / 100f,
-        animationSpec = tween(MotionMedium, easing = MotionStandard),
-        label = "readonlyProgress",
-    )
-    Box(
-        modifier = modifier
-            .height(12.dp)
-            .clip(RoundedCornerShape(6.dp))
-            .background(WarmLine.copy(alpha = 0.68f)),
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxHeight()
-                .fillMaxWidth(animated)
-                .background(WarmBrown.copy(alpha = 0.68f)),
-        )
-    }
-}
-
-@Composable
-private fun ReadOnlyPriorityBar(priority: Int, modifier: Modifier = Modifier) {
-    val normalizedPriority = priority.coerceIn(1, 9)
-    val thumbFraction = (9 - normalizedPriority) / 8f
-    val trackColor = WarmLine
-    Canvas(
-        modifier = modifier
-            .height(34.dp)
-            .padding(vertical = 8.dp),
-    ) {
-        val barHeight = 12.dp.toPx()
-        val radius = barHeight / 2f
-        val centerY = size.height / 2f
-        // Inset the dots by the bar's corner radius so the first/last dot sit inside the rounded
-        // caps rather than poking past them.
-        val dotSpanStart = radius
-        val dotSpanWidth = (size.width - radius * 2f).coerceAtLeast(1f)
-        val fillEnd = dotSpanStart + dotSpanWidth * thumbFraction
-        drawRoundRect(
-            color = trackColor.copy(alpha = 0.62f),
-            topLeft = Offset(0f, centerY - barHeight / 2f),
-            size = Size(size.width, barHeight),
-            cornerRadius = CornerRadius(radius, radius),
-        )
-        (1..9).forEach { level ->
-            val x = dotSpanStart + dotSpanWidth * ((9 - level) / 8f)
-            drawCircle(
-                color = priorityColor(level),
-                radius = 4.dp.toPx(),
-                center = Offset(x, centerY),
-            )
-        }
-        // Solid active fill drawn OVER the lower-priority (higher-number) dots, like the editor
-        // slider — replaces the old faint fill that sat under the dots.
-        drawRoundRect(
-            color = priorityColor(normalizedPriority),
-            topLeft = Offset(0f, centerY - barHeight / 2f),
-            size = Size(fillEnd, barHeight),
-            cornerRadius = CornerRadius(radius, radius),
-        )
-        // Current-value marker on top of the fill so it stays visible at the fill's leading edge.
-        drawCircle(color = Color.White, radius = 5.5.dp.toPx(), center = Offset(fillEnd, centerY))
-        drawCircle(color = priorityColor(normalizedPriority), radius = 3.6.dp.toPx(), center = Offset(fillEnd, centerY))
-    }
-}
-
-@Composable
-private fun LocationDetailLine(location: String, mapVerified: Boolean?, autoLoadMapPreviews: Boolean) {
-    val context = LocalContext.current
-    val displayValue = remember(location) { location.cleanCalendarDisplayText() }
-    Surface(
-        shape = RoundedCornerShape(14.dp),
-        color = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.38f),
-        modifier = Modifier
-            .fillMaxWidth()
-            .animateContentSize(animationSpec = tween(MotionMedium, easing = MotionStandard)),
-    ) {
-        Column(
-            modifier = Modifier.padding(horizontal = 14.dp, vertical = 11.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            Text(stringResource(R.string.location), color = MaterialTheme.colorScheme.onSurfaceVariant, fontSize = 12.sp, lineHeight = 14.sp, fontWeight = FontWeight.SemiBold)
-            LinkableSelectableText(displayValue, onClick = { context.openMapLocation(displayValue) })
-            LocationMapPreview(displayValue, mapVerified, autoLoadMapPreviews)
-        }
-    }
-}
-
-@Composable
-private fun LocationMapPreview(location: String, mapVerified: Boolean?, autoLoadMapPreviews: Boolean) {
-    if (!location.shouldAttemptMapPreview(mapVerified)) return
-    var suggestion by remember(location) { mutableStateOf<LocationSuggestion?>(null) }
-    var previewRequested by remember(location, autoLoadMapPreviews) { mutableStateOf(autoLoadMapPreviews) }
-    var resolved by remember(location) { mutableStateOf(false) }
-
-    if (!previewRequested) {
-        Column(verticalArrangement = Arrangement.spacedBy(6.dp)) {
-            OutlinedButton(
-                onClick = { previewRequested = true },
-                modifier = Modifier.fillMaxWidth(),
-            ) {
-                Icon(Icons.Default.LocationOn, contentDescription = null, modifier = Modifier.size(18.dp))
-                Spacer(Modifier.width(8.dp))
-                Text(stringResource(R.string.show_map_preview))
-            }
-            Text(
-                stringResource(R.string.osm_attribution),
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                fontSize = 11.sp,
-                lineHeight = 13.sp,
-                modifier = Modifier.padding(horizontal = 4.dp),
-            )
-        }
-        return
-    }
-
-    LaunchedEffect(location, previewRequested) {
-        if (previewRequested) {
-            resolved = false
-            suggestion = runCatching { LocationLookup.search(location, limit = 1, allowAliases = false).firstOrNull() }.getOrNull()
-            resolved = true
-        }
-    }
-
-    val resolvedSuggestion = suggestion ?: return
-    Surface(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(128.dp),
-        shape = RoundedCornerShape(16.dp),
-        color = WarmLine.copy(alpha = if (resolved) 0.35f else 0.18f),
-        border = BorderStroke(1.dp, WarmLine.copy(alpha = 0.7f)),
-    ) {
-        OpenStreetMapPreview(
-            suggestion = resolvedSuggestion,
-            modifier = Modifier.fillMaxSize(),
-        )
-    }
-}
-
-@Composable
-private fun OpenStreetMapPreview(suggestion: LocationSuggestion, modifier: Modifier = Modifier) {
-    val context = LocalContext.current
-    val mapModifier = modifier.clickable { context.openMapLocation(suggestion) }
-    var imageBitmap by remember(suggestion.latitude, suggestion.longitude) { mutableStateOf<androidx.compose.ui.graphics.ImageBitmap?>(null) }
-    var failed by remember(suggestion.latitude, suggestion.longitude) { mutableStateOf(false) }
-    LaunchedEffect(suggestion.latitude, suggestion.longitude) {
-        failed = false
-        imageBitmap = runCatching {
-            withContext(Dispatchers.IO) {
-                LocationLookup.mapPreviewBitmap(suggestion).asImageBitmap()
-            }
-        }.getOrNull()
-        failed = imageBitmap == null
-    }
-    Box(mapModifier.clipToBounds()) {
-        when {
-            imageBitmap != null -> Image(
-                bitmap = imageBitmap!!,
-                contentDescription = stringResource(R.string.map_preview_content_description),
-                contentScale = ContentScale.Crop,
-                modifier = Modifier.fillMaxSize(),
-            )
-            failed -> Box(Modifier.fillMaxSize().background(WarmLine.copy(alpha = 0.32f)))
-            else -> Box(Modifier.fillMaxSize().background(WarmLine.copy(alpha = 0.18f)))
-        }
-        Surface(
-            modifier = Modifier
-                .align(Alignment.BottomEnd)
-                .padding(6.dp),
-            color = MaterialTheme.colorScheme.surface.copy(alpha = 0.86f),
-            shape = RoundedCornerShape(8.dp),
-        ) {
-            Text(
-                stringResource(R.string.osm_attribution),
-                color = WarmInk,
-                fontSize = 9.sp,
-                lineHeight = 11.sp,
-                modifier = Modifier.padding(horizontal = 6.dp, vertical = 3.dp),
-            )
-        }
-    }
-}
-
-@Composable
-private fun LinkableSelectableText(value: String, onClick: (() -> Unit)? = null) {
-    val textColor = WarmInk
-    val linkColor = MaterialTheme.colorScheme.primary
-    val highlightColor = WarmPeach.copy(alpha = 0.64f).toArgb()
-    AndroidView(
-        modifier = Modifier.fillMaxWidth(),
-        factory = { context ->
-            TextView(context).apply {
-                includeFontPadding = false
-                setTextIsSelectable(true)
-                linksClickable = true
-                movementMethod = LinkMovementMethod.getInstance()
-                setTextSize(TypedValue.COMPLEX_UNIT_SP, 16f)
-                setLineSpacing(0f, 1.08f)
-                this.highlightColor = highlightColor
-            }
-        },
-        update = { textView ->
-            val linkedText = SpannableString(value)
-            Linkify.addLinks(linkedText, Linkify.WEB_URLS or Linkify.EMAIL_ADDRESSES)
-            textView.text = linkedText
-            textView.setTextColor(textColor.toArgb())
-            textView.setLinkTextColor(linkColor.toArgb())
-            textView.isClickable = onClick != null
-            textView.setOnClickListener(if (onClick != null) android.view.View.OnClickListener { onClick() } else null)
-        },
-    )
 }
 
 @Composable
@@ -6678,28 +4791,6 @@ private fun InactiveCollectionOrderRow(
     }
 }
 
-private enum class SettingsDestination(val title: String) {
-    Main("Settings"),
-    Accounts("Calendar"),
-    AddSource("Add calendar"),
-    AddAccount("Add CalDAV"),
-    AddReadOnly("Read-only calendar"),
-    AccountDetail("Edit source"),
-    Behavior("Behavior"),
-    Design("Design"),
-    Widgets("Widgets"),
-    WidgetAgenda("Agenda widget"),
-    WidgetMonth("Month widget"),
-    WidgetTasks("Tasks widget"),
-    WidgetMulti("Multi widget"),
-    WidgetDay("Day widget"),
-    Privacy("Privacy"),
-    EventFieldOrder("Event fields"),
-    TaskFieldOrder("Task fields"),
-    Sources("Calendars & sources"),
-    Reorder("Order"),
-}
-
 @Composable
 private fun SettingsDestination.localizedTitle(): String = when (this) {
     SettingsDestination.Main -> stringResource(R.string.settings)
@@ -7872,7 +5963,7 @@ internal object LocationLookup {
     }
 }
 
-private fun String.shouldAttemptMapPreview(mapVerified: Boolean?): Boolean {
+internal fun String.shouldAttemptMapPreview(mapVerified: Boolean?): Boolean {
     if (mapVerified == false) return false
     if (mapVerified == true) return cleanCalendarDisplayText().length >= 3
     val cleaned = cleanCalendarDisplayText()
@@ -7882,7 +5973,7 @@ private fun String.shouldAttemptMapPreview(mapVerified: Boolean?): Boolean {
     return hasAddressSignal && !cleaned.contains('\n')
 }
 
-private fun Context.openMapLocation(suggestion: LocationSuggestion) {
+internal fun Context.openMapLocation(suggestion: LocationSuggestion) {
     val latLng = "${suggestion.latitude},${suggestion.longitude}"
     val encodedLabel = Uri.encode(suggestion.primaryName.ifBlank { suggestion.displayName })
     val mapIntent = Intent(Intent.ACTION_VIEW, Uri.parse("geo:${suggestion.latitude},${suggestion.longitude}?q=$latLng($encodedLabel)"))
@@ -7891,7 +5982,7 @@ private fun Context.openMapLocation(suggestion: LocationSuggestion) {
         .recoverCatching { startActivity(fallbackIntent) }
 }
 
-private fun Context.openMapLocation(location: String) {
+internal fun Context.openMapLocation(location: String) {
     val cleaned = location.cleanCalendarDisplayText()
     if (cleaned.isBlank()) return
     val encodedQuery = Uri.encode(cleaned)
@@ -7901,7 +5992,7 @@ private fun Context.openMapLocation(location: String) {
         .recoverCatching { startActivity(fallbackIntent) }
 }
 
-private fun TaskEntity.displayProgress(): Int =
+internal fun TaskEntity.displayProgress(): Int =
     when {
         isCompleted -> 100
         percentComplete != null -> percentComplete.coerceIn(0, 100)
@@ -8040,7 +6131,7 @@ private fun AppLanguageMode.localizedLabel(): String = when (this) {
 }
 
 @Composable
-private fun TaskEntity.localizedStatusLabel(): String = when (status?.uppercase()) {
+internal fun TaskEntity.localizedStatusLabel(): String = when (status?.uppercase()) {
     "IN-PROCESS" -> appString(R.string.in_progress)
     "CANCELLED" -> appString(R.string.aborted)
     "COMPLETED" -> appString(R.string.status_completed)
@@ -8214,7 +6305,7 @@ private fun <T> rememberSmoothRemoval(
 }
 
 @Composable
-private fun pendingMutationFor(resourceHref: String): PendingMutationEntity? {
+internal fun pendingMutationFor(resourceHref: String): PendingMutationEntity? {
     val mutations = LocalPendingMutations.current
     return remember(mutations, resourceHref) {
         mutations.lastOrNull { it.resourceHref == resourceHref }
@@ -8304,540 +6395,6 @@ internal fun ParticipantRoleOption.localizedLabel(): String = when (this) {
     ParticipantRoleOption.NonParticipant -> appString(R.string.non_participant)
 }
 
-@Composable
-internal fun String.localizedParticipantRoleLabel(): String =
-    ParticipantRoleOption.entries.firstOrNull { it.value.equals(this, ignoreCase = true) }?.localizedLabel() ?: this
-
-@Composable
-internal fun EventParticipantStack(
-    attendees: List<CalendarParticipant>,
-    eventColor: Color,
-    contentColor: Color,
-    modifier: Modifier = Modifier,
-    circleSize: Dp = 26.dp,
-    maxVisible: Int = 5,
-    muted: Boolean = false,
-) {
-    if (attendees.isEmpty() || maxVisible <= 0) return
-    BoxWithConstraints(
-        modifier = modifier
-            .height(circleSize + 12.dp)
-            .graphicsLayer { clip = false },
-        contentAlignment = Alignment.CenterEnd,
-    ) {
-        val step = circleSize * 0.64f
-        val horizontalPadding = 4.dp
-        val verticalPadding = 3.dp
-        val byWidth = (((maxWidth - horizontalPadding * 2 - circleSize) / step).toInt() + 1).coerceAtLeast(1)
-        val capacity = min(maxVisible, byWidth)
-        val needsMore = attendees.size > capacity
-        val visibleCount = if (needsMore) (capacity - 1).coerceAtLeast(1) else capacity
-        val visible = attendees.take(visibleCount)
-        val totalCount = visible.size + if (needsMore) 1 else 0
-        val stackWidth = circleSize + step * (totalCount - 1).coerceAtLeast(0).toFloat()
-        val containerColor = if (muted) {
-            MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.72f)
-        } else {
-            Color.White.copy(alpha = 0.94f)
-        }
-        Box(
-            modifier = Modifier
-                .width(stackWidth + horizontalPadding * 2)
-                .height(circleSize + verticalPadding * 2)
-                .graphicsLayer { clip = false },
-        ) {
-            val containerShape = RoundedCornerShape(999.dp)
-            Box(
-                Modifier
-                    .matchParentSize()
-                    .clip(containerShape)
-                    .background(containerColor),
-            )
-            Box(
-                modifier = Modifier
-                    .matchParentSize()
-                    .graphicsLayer { clip = false },
-            ) {
-                visible.forEachIndexed { index, attendee ->
-                    ParticipantCircle(
-                        label = attendee.initials(),
-                        color = attendee.avatarColor().participantMuted(muted, 0.70f),
-                        contentColor = if (muted) WarmInk.copy(alpha = 0.66f) else Color.White,
-                        borderColor = attendee.displayParticipationColor().participantMuted(muted, 0.82f),
-                        modifier = Modifier
-                            .offset(x = horizontalPadding + step * index.toFloat(), y = verticalPadding)
-                            .size(circleSize),
-                    )
-                }
-                if (needsMore) {
-                    ParticipantCircle(
-                        label = "+${attendees.size - visible.size}",
-                        color = contentColor.participantMuted(muted, 0.78f).copy(alpha = 0.92f),
-                        contentColor = if (muted) WarmInk.copy(alpha = 0.66f) else if (contentColor.isDark()) Color.White else Color(0xFF1C1A18),
-                        borderColor = eventColor.participantMuted(muted, 0.82f),
-                        modifier = Modifier
-                            .offset(x = horizontalPadding + step * visible.size.toFloat(), y = verticalPadding)
-                            .size(circleSize),
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-internal fun CompactParticipantStatusDots(
-    attendees: List<CalendarParticipant>,
-    modifier: Modifier = Modifier,
-    maxVisible: Int = 3,
-    scale: Float = 1f,
-) {
-    val visible = attendees.take(maxVisible)
-    val hasMore = attendees.size > visible.size
-    val safeScale = scale.coerceIn(0.34f, 1f)
-    val dotSize = (5f * safeScale).dp
-    val horizontalPadding = (5f * safeScale).dp
-    val verticalPadding = (2.5f * safeScale).dp
-    val spacing = (3f * safeScale).dp
-    Surface(
-        modifier = modifier,
-        shape = RoundedCornerShape(topStart = 8.dp, topEnd = 8.dp, bottomStart = 2.dp, bottomEnd = 2.dp),
-        color = Color.White.copy(alpha = 0.94f),
-    ) {
-        Row(
-            modifier = Modifier.padding(horizontal = horizontalPadding, vertical = verticalPadding),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(spacing),
-        ) {
-            visible.forEach { attendee ->
-                Box(
-                    modifier = Modifier
-                        .size(dotSize)
-                        .clip(CircleShape)
-                        .background(attendee.displayParticipationColor()),
-                )
-            }
-            if (hasMore) {
-                Box(
-                    modifier = Modifier
-                        .size(dotSize)
-                        .clip(CircleShape)
-                        .background(MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.62f)),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun ParticipantCircle(
-    label: String,
-    color: Color,
-    contentColor: Color,
-    borderColor: Color,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        modifier = modifier
-            .clip(CircleShape)
-            .background(color)
-            .border(2.5.dp, borderColor.copy(alpha = 0.96f), CircleShape),
-        contentAlignment = Alignment.Center,
-    ) {
-        Text(
-            label,
-            color = contentColor,
-            fontSize = 9.sp,
-            lineHeight = 10.sp,
-            fontWeight = FontWeight.SemiBold,
-            maxLines = 1,
-        )
-    }
-}
-
-private fun CalendarParticipant.initials(): String {
-    val source = displayName.ifBlank { email }
-    val parts = source.split(Regex("\\s+")).filter { it.isNotBlank() }
-    val initials = when {
-        parts.size >= 2 -> "${parts.first().first()}${parts.last().first()}"
-        source.isNotBlank() -> source.take(2)
-        else -> "?"
-    }
-    return initials.uppercase(Locale.ROOT)
-}
-
-private fun CalendarParticipant.avatarColor(): Color {
-    val colors = listOf(
-        Color(0xFF2563A8),
-        Color(0xFF0F766E),
-        Color(0xFF7C3AED),
-        Color(0xFFC2410C),
-        Color(0xFFBE123C),
-        Color(0xFF4D7C0F),
-    )
-    val index = abs((email.ifBlank { displayName }).hashCode()) % colors.size
-    return colors[index]
-}
-
-private fun Color.participantMuted(muted: Boolean, alpha: Float): Color =
-    if (muted) greyedOut(0.78f).copy(alpha = alpha) else this
-
-internal fun String?.toCalendarParticipant(): CalendarParticipant? =
-    runCatching {
-        if (isNullOrBlank()) return@runCatching null
-        val obj = JSONObject(this)
-        val email = obj.optString("email")
-        CalendarParticipant(
-            name = obj.optString("name", email),
-            email = email,
-            partstat = obj.optString("partstat", "NEEDS-ACTION"),
-            role = obj.optString("role", "REQ-PARTICIPANT"),
-            rsvp = obj.optString("rsvp").equals("TRUE", ignoreCase = true),
-            scheduleStatus = obj.optString("scheduleStatus").takeIf { it.isNotBlank() },
-        )
-    }.getOrNull()
-
-internal fun String?.toCalendarParticipants(): List<CalendarParticipant> =
-    runCatching {
-        if (isNullOrBlank()) return@runCatching emptyList()
-        val array = JSONArray(this)
-        buildList {
-            repeat(array.length()) { index ->
-                val obj = array.optJSONObject(index) ?: return@repeat
-                val email = obj.optString("email")
-                if (email.isBlank()) return@repeat
-                add(
-                    CalendarParticipant(
-                        name = obj.optString("name", email),
-                        email = email,
-                        partstat = obj.optString("partstat", "NEEDS-ACTION"),
-                        role = obj.optString("role", "REQ-PARTICIPANT"),
-                        rsvp = obj.optString("rsvp").equals("TRUE", ignoreCase = true),
-                        scheduleStatus = obj.optString("scheduleStatus").takeIf { it.isNotBlank() },
-                    ),
-                )
-            }
-        }
-    }.getOrDefault(emptyList())
-
-internal fun List<CalendarParticipant>.toAttendeesJson(): String? =
-    takeIf { it.isNotEmpty() }?.let { participants ->
-        JSONArray().apply {
-            participants.forEach { attendee ->
-                put(
-                    JSONObject().apply {
-                        put("name", attendee.name)
-                        put("email", attendee.email)
-                        put("partstat", attendee.partstat)
-                        put("role", attendee.role)
-                        put("rsvp", if (attendee.rsvp) "TRUE" else "FALSE")
-                        attendee.scheduleStatus?.let { put("scheduleStatus", it) }
-                    },
-                )
-            }
-        }.toString()
-    }
-
-internal fun String.isLikelyEmailAddress(): Boolean =
-    matches(Regex("^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$"))
-
-private fun String.participantRoleLabel(): String =
-    ParticipantRoleOption.entries.firstOrNull { it.value.equals(this, ignoreCase = true) }?.label ?: this
-
-private fun String.eventStatusLabel(): String = when (uppercase(Locale.US)) {
-    "TENTATIVE" -> "Tentative"
-    "CANCELLED" -> "Cancelled"
-    else -> "Confirmed"
-}
-
-@Composable
-private fun EventEntity.statusLabelText(): String = when ((status ?: EventStatusOption.Confirmed.value).uppercase(Locale.US)) {
-    "TENTATIVE" -> appString(R.string.tentative)
-    "CANCELLED" -> appString(R.string.cancelled)
-    else -> appString(R.string.confirmed)
-}
-
-private fun String.eventClassLabel(): String = when (uppercase(Locale.US)) {
-    "CONFIDENTIAL" -> "Show busy status when shared"
-    "PRIVATE" -> "Hide this event when shared"
-    else -> "Show full event when shared"
-}
-
-@Composable
-private fun EventEntity.classLabelText(): String = when ((classification ?: EventClassOption.Public.value).uppercase(Locale.US)) {
-    "CONFIDENTIAL" -> appString(R.string.busy_only)
-    "PRIVATE" -> appString(R.string.hidden)
-    else -> appString(R.string.full_details)
-}
-
-private fun String.eventTransparencyLabel(): String = when (uppercase(Locale.US)) {
-    "TRANSPARENT" -> "Free"
-    else -> "Busy"
-}
-
-@Composable
-private fun EventEntity.transparencyLabelText(): String = when ((transparency ?: EventTransparencyOption.Busy.value).uppercase(Locale.US)) {
-    "TRANSPARENT" -> appString(R.string.free)
-    else -> appString(R.string.busy)
-}
-
-private fun String.participationLabel(): String = when (uppercase(Locale.US)) {
-    "ACCEPTED" -> "Accepted"
-    "DECLINED" -> "Declined"
-    "TENTATIVE" -> "Tentative"
-    "DELEGATED" -> "Delegated"
-    else -> "No response"
-}
-
-@Composable
-private fun String.participationColor(): Color = when (uppercase(Locale.US)) {
-    "ACCEPTED" -> Color(0xFF00A86B)
-    "DECLINED" -> Color(0xFFE53935)
-    "TENTATIVE" -> Color(0xFFFFA000)
-    "DELEGATED" -> Color(0xFF5B6CFF)
-    else -> Color(0xFF6B7280)
-}
-
-private fun CalendarParticipant.displayParticipationLabel(): String =
-    deliveryStatusLabel() ?: partstat.participationLabel()
-
-@Composable
-private fun CalendarParticipant.localizedDisplayParticipationLabel(): String =
-    localizedDeliveryStatusLabel() ?: when (partstat.uppercase(Locale.US)) {
-        "ACCEPTED" -> appString(R.string.accept)
-        "DECLINED" -> appString(R.string.decline)
-        "TENTATIVE" -> appString(R.string.tentative)
-        "DELEGATED" -> appString(R.string.delegated)
-        else -> appString(R.string.no_response)
-    }
-
-@Composable
-private fun CalendarParticipant.displayParticipationColor(): Color =
-    if (deliveryStatusLabel() != null) deliveryStatusColor() else partstat.participationColor()
-
-private fun CalendarParticipant.deliveryStatusLabel(): String? {
-    val code = scheduleStatus
-        ?.substringBefore(',')
-        ?.substringBefore(';')
-        ?.trim()
-        ?.takeIf { it.isNotBlank() }
-        ?: return null
-    return when {
-        code == "1.0" -> "Preparing invitation"
-        code == "1.1" -> "Invitation sent"
-        code == "1.2" || code == "2.0" -> null
-        code == "3.7" -> "Invitation could not be delivered"
-        code.startsWith("3.") -> "Invitation could not be delivered"
-        code.startsWith("5.") -> "Invitation could not be delivered"
-        else -> null
-    }
-}
-
-@Composable
-internal fun CalendarParticipant.deliveryStatusColor(): Color =
-    when {
-        scheduleStatus.orEmpty().startsWith("1.") -> Color(0xFF4DA3FF)
-        else -> Color(0xFFE53935)
-    }
-
-private val TaskStatusOptions = listOf(
-    "NEEDS-ACTION",
-    "IN-PROCESS",
-    "COMPLETED",
-    "CANCELLED",
-)
-
-@Composable
-private fun taskStatusOptionLabel(value: String): String = when (value) {
-    "IN-PROCESS" -> appString(R.string.in_progress)
-    "COMPLETED" -> appString(R.string.status_completed)
-    "CANCELLED" -> appString(R.string.aborted)
-    else -> appString(R.string.status_open)
-}
-
-/** Normalised effective status, falling back to the legacy isCompleted boolean. */
-internal fun TaskEntity.effectiveStatus(): String = status?.uppercase()
-    ?: if (isCompleted) "COMPLETED" else "NEEDS-ACTION"
-
-/**
- * A task is "inactive" when it's done OR cancelled — both should be greyed out and have
- * their priority animation suppressed.
- */
-internal fun TaskEntity.isInactive(): Boolean = isCompleted || effectiveStatus() == "CANCELLED"
-
-/** Sort weight for the "Status" sort: In Bearbeitung first, then Offen, then others. */
-internal fun TaskEntity.statusSortRank(): Int = when (effectiveStatus()) {
-    "IN-PROCESS" -> 0
-    "NEEDS-ACTION" -> 1
-    "COMPLETED" -> 2
-    "CANCELLED" -> 3
-    else -> 4
-}
-
-internal enum class PlannedTaskSort(val label: String) {
-    Date("Date"),
-    Priority("Priority"),
-    Status("Status");
-
-    fun next(): PlannedTaskSort = entries[(ordinal + 1) % entries.size]
-}
-
-@Composable
-internal fun PlannedTaskSort.localizedLabel(): String = when (this) {
-    PlannedTaskSort.Date -> appString(R.string.date)
-    PlannedTaskSort.Priority -> appString(R.string.priority)
-    PlannedTaskSort.Status -> appString(R.string.status)
-}
-
-/**
- * Shared task checkbox used in the day view, sidebar, detail sheet and editor.
- *
- * - Tap toggles between COMPLETED and NEEDS-ACTION, unless the caller requests the picker.
- * - Long-press opens a status menu (Offen / In Bearbeitung / Erledigt / Abgebrochen).
- * - The glyph reflects the status: empty circle (open), slow indeterminate spinner
- *   (in progress), filled check (completed) and an X (cancelled).
- */
-@Composable
-internal fun TaskStatusCheckbox(
-    status: String,
-    tint: Color,
-    onStatusChange: (String) -> Unit,
-    modifier: Modifier = Modifier,
-    tapOpensPicker: Boolean = false,
-    boxSize: Dp = 40.dp,
-    iconSize: Dp = 24.dp,
-) {
-    var menuOpen by remember { mutableStateOf(false) }
-    Box(
-        modifier = modifier
-            .size(boxSize)
-            .clip(CircleShape)
-            .combinedClickable(
-                onClick = {
-                    if (tapOpensPicker) {
-                        menuOpen = true
-                    } else {
-                        onStatusChange(if (status == "COMPLETED") "NEEDS-ACTION" else "COMPLETED")
-                    }
-                },
-                onLongClick = { menuOpen = true },
-            ),
-        contentAlignment = Alignment.Center,
-    ) {
-        // The outer ring is always drawn; the state glyph (check / X / spinner) lives
-        // *inside* the ring, never replacing it.
-        if (status == "COMPLETED") {
-            Icon(
-                Icons.Default.CheckCircle,
-                contentDescription = null,
-                tint = tint,
-                modifier = Modifier.size(iconSize),
-            )
-        } else {
-            Icon(
-                Icons.Default.RadioButtonUnchecked,
-                contentDescription = null,
-                tint = tint,
-                modifier = Modifier.size(iconSize),
-            )
-            // Inner glyph centered inside the ring.
-            when (status) {
-                "IN-PROCESS" -> Box(
-                    modifier = Modifier
-                        .size(iconSize * 0.34f)
-                        .clip(CircleShape)
-                        .background(tint),
-                )
-                "CANCELLED" -> Icon(
-                    Icons.Default.Close,
-                    contentDescription = null,
-                    tint = tint,
-                    modifier = Modifier.size(iconSize * 0.58f),
-                )
-            }
-        }
-    }
-    if (menuOpen) {
-        TaskStatusPickerDialog(
-            current = status,
-            onPick = {
-                menuOpen = false
-                onStatusChange(it)
-            },
-            onDismiss = { menuOpen = false },
-        )
-    }
-}
-
-/**
- * Centered, rounded status picker. Rendered as a Dialog (its own window) so it is
- * never affected by the card's priority wiggle animation, and its corners match the
- * rounded look of the card overflow menu.
- */
-@Composable
-private fun TaskStatusPickerDialog(
-    current: String,
-    onPick: (String) -> Unit,
-    onDismiss: () -> Unit,
-) {
-    Dialog(onDismissRequest = onDismiss) {
-        Surface(
-            shape = RoundedCornerShape(28.dp),
-            color = MaterialTheme.colorScheme.surface,
-            tonalElevation = 3.dp,
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            Column(modifier = Modifier.padding(vertical = 10.dp)) {
-                Text(
-                    appString(R.string.status),
-                    color = WarmInk,
-                    fontSize = 14.sp,
-                    lineHeight = 17.sp,
-                    fontWeight = FontWeight.SemiBold,
-                    modifier = Modifier.padding(horizontal = 22.dp, vertical = 8.dp),
-                )
-                TaskStatusOptions.forEach { value ->
-                    val label = taskStatusOptionLabel(value)
-                    val selected = value == current
-                    Row(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable { onPick(value) }
-                            .padding(horizontal = 22.dp, vertical = 14.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(12.dp),
-                    ) {
-                        Box(Modifier.size(22.dp), contentAlignment = Alignment.Center) {
-                            Icon(Icons.Default.RadioButtonUnchecked, contentDescription = null, tint = WarmInk, modifier = Modifier.size(22.dp))
-                            when (value) {
-                                "IN-PROCESS" -> Box(Modifier.size(8.dp).clip(CircleShape).background(WarmInk))
-                                "CANCELLED" -> Icon(Icons.Default.Close, contentDescription = null, tint = WarmInk, modifier = Modifier.size(13.dp))
-                                "COMPLETED" -> Icon(Icons.Default.Check, contentDescription = null, tint = WarmInk, modifier = Modifier.size(13.dp))
-                            }
-                        }
-                        Text(
-                            label,
-                            color = WarmInk,
-                            fontSize = 16.sp,
-                            lineHeight = 19.sp,
-                            fontWeight = if (selected) FontWeight.Bold else FontWeight.Normal,
-                            modifier = Modifier.weight(1f),
-                        )
-                        if (selected) {
-                            Icon(Icons.Default.Check, contentDescription = null, tint = WarmBrown, modifier = Modifier.size(20.dp))
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * Returns the list with the calendar matching [defaultHref] (if any) pulled to the
- * front, keeping the relative order of the remaining items. Used wherever the user
- * picks a calendar for a new entry so the chosen default is always the first chip.
- */
 internal fun List<CollectionEntity>.sortedWithDefaultFirst(defaultHref: String?): List<CollectionEntity> {
     if (defaultHref.isNullOrBlank()) return this
     val default = firstOrNull { it.href == defaultHref } ?: return this
@@ -8853,9 +6410,9 @@ private fun CollectionEntity.canDeleteFromServerForUi(): Boolean =
         !readOnly &&
         capabilitiesJson.toJsonObjectOrNull()?.optBooleanOrNull("canDeleteResources") != false
 
-private fun String.isLocalCollectionHrefUi(): Boolean = startsWith(UiLocalCollectionPrefix)
+internal fun String.isLocalCollectionHrefUi(): Boolean = startsWith(UiLocalCollectionPrefix)
 
-private fun CollectionEntity.isAndroidProviderForUi(): Boolean =
+internal fun CollectionEntity.isAndroidProviderForUi(): Boolean =
     sourceType == SourceType.AndroidProvider || href.startsWith(UiAndroidCollectionPrefix)
 
 internal fun AccountEntity.isAndroidProviderForUi(): Boolean =
@@ -9061,7 +6618,7 @@ internal fun TaskEntity.displayColor(mode: TaskColorMode): Int =
         TaskColorMode.Priority -> priority?.let { priorityColor(it).toArgb() } ?: color
     }
 
-private fun String.cleanCalendarDisplayText(): String {
+internal fun String.cleanCalendarDisplayText(): String {
     var normalized = trim().trimHtmlDataPrefixForDisplay()
     if (normalized.contains(Regex("%[0-9A-Fa-f]{2}"))) {
         normalized = runCatching { Uri.decode(normalized) }.getOrDefault(normalized).trimHtmlDataPrefixForDisplay()
@@ -9165,7 +6722,7 @@ internal fun String.toLocalizedRecurrenceLabel(): String {
 }
 
 @Composable
-private fun String?.localizedReminderSummary(): String? {
+internal fun String?.localizedReminderSummary(): String? {
     val minutes = this?.split(',')?.mapNotNull { it.trim().toIntOrNull() }?.normalizedReminderOffsets().orEmpty()
     if (minutes.isEmpty()) return null
     return minutes.map { reminderMinuteLabel(it) }.joinToString(", ")
@@ -9482,38 +7039,6 @@ private fun String.toDisplayDate(): String = runCatching {
     LocalDate.parse(this).format(DateTimeFormatter.ofPattern("EEE, d. MMM yyyy", Locale.getDefault()))
 }.getOrDefault(this)
 
-sealed interface CreationSheet {
-    data object EventLow : CreationSheet
-    data object EventFull : CreationSheet
-    data object TaskLow : CreationSheet
-    data object Task : CreationSheet
-    data class TaskForParent(val parent: TaskEntity) : CreationSheet
-    data class EditEvent(val event: EventEntity) : CreationSheet
-    data class EditTask(val task: TaskEntity) : CreationSheet
-    data class DuplicateEvent(val event: EventEntity) : CreationSheet
-    data class DuplicateTask(val task: TaskEntity) : CreationSheet
-}
-
-private data class HiddenSaveNotice(
-    val collectionHref: String,
-    val kind: HiddenSaveKind,
-)
-
-private enum class HiddenSaveKind {
-    Event,
-    Task,
-}
-
-private sealed interface ConversionSource {
-    data class Event(val event: EventEntity) : ConversionSource
-    data class Task(val task: TaskEntity) : ConversionSource
-}
-
-private sealed interface RecurringSaveRequest {
-    data class Event(val event: EventEntity, val payload: EventEditPayload) : RecurringSaveRequest
-    data class Task(val task: TaskEntity, val payload: TaskEditPayload) : RecurringSaveRequest
-}
-
 internal enum class SheetSnap {
     Expanded,
     Half,
@@ -9525,116 +7050,6 @@ internal enum class SheetSnap {
 internal enum class SheetAnchorMode {
     ContentFit,
     Editor,
-}
-
-private fun DetailSheet.preferredInitialSnap(): SheetSnap = when (this) {
-    is DetailSheet.Event -> {
-        val description = event.description.orEmpty().cleanCalendarDisplayText()
-        val metadataScore =
-            (if (description.length >= 72 || description.count { it == '\n' } >= 2) 2 else if (description.isNotBlank()) 1 else 0) +
-                (if (event.location.orEmpty().isNotBlank()) 1 else 0) +
-                (if (event.categories.orEmpty().isNotBlank()) 1 else 0) +
-                (if (event.attendeesJson.orEmpty().isNotBlank()) 2 else 0) +
-                (if (!event.status.isNullOrBlank() || !event.classification.isNullOrBlank() || !event.transparency.isNullOrBlank()) 1 else 0) +
-                (if (!event.recurrenceRule.isNullOrBlank()) 1 else 0) +
-                (if (event.title.length >= 34) 1 else 0)
-        when {
-            metadataScore >= 2 -> SheetSnap.Expanded
-            metadataScore == 0 -> SheetSnap.Quarter
-            else -> SheetSnap.Half
-        }
-    }
-    is DetailSheet.Task -> {
-        val notes = task.notes.orEmpty().cleanCalendarDisplayText()
-        val metadataScore =
-            (if (notes.length >= 72 || notes.count { it == '\n' } >= 2) 2 else if (notes.isNotBlank()) 1 else 0) +
-                (if (task.location.orEmpty().isNotBlank()) 1 else 0) +
-                (if (task.url.orEmpty().isNotBlank()) 1 else 0) +
-                (if (task.categories.orEmpty().isNotBlank()) 1 else 0) +
-                (if (!task.recurrenceRule.isNullOrBlank()) 1 else 0) +
-                (if (task.priority != null && task.priority != 9) 1 else 0) +
-                (if (task.displayProgress() > 0) 1 else 0) +
-                (if (task.title.length >= 34) 1 else 0)
-        when {
-            metadataScore >= 2 -> SheetSnap.Expanded
-            metadataScore == 0 -> SheetSnap.Quarter
-            else -> SheetSnap.Half
-        }
-    }
-}
-
-private fun DetailSheet.estimatedPopoverHeight(): Dp = when (this) {
-    is DetailSheet.Event -> {
-        val description = event.description.orEmpty().cleanCalendarDisplayText()
-        val location = event.location.orEmpty()
-        val height = 186 +
-            82 +
-            detailTextBlockHeight(description, baseWhenBlank = 0) +
-            if (location.isNotBlank()) {
-                86 + if (location.shouldAttemptMapPreview(event.locationMapVerified)) 150 else 0
-            } else {
-                0
-            } +
-            if (!event.categories.isNullOrBlank()) 72 else 0 +
-            event.estimatedParticipantsDetailHeight() +
-            120 +
-            if (!event.recurrenceRule.isNullOrBlank()) 72 else 0 +
-            if (!event.syncError.isNullOrBlank()) 72 else 0 +
-            82
-        height.dp
-    }
-    is DetailSheet.Task -> {
-        val notes = task.notes.orEmpty().cleanCalendarDisplayText()
-        val location = task.location.orEmpty()
-        val height = 186 +
-            72 +
-            76 +
-            76 +
-            if (!task.recurrenceRule.isNullOrBlank()) 72 else 0 +
-            detailTextBlockHeight(notes, baseWhenBlank = 0) +
-            if (location.isNotBlank()) {
-                86 + if (location.shouldAttemptMapPreview(task.locationMapVerified)) 150 else 0
-            } else {
-                0
-            } +
-            if (!task.url.isNullOrBlank()) 72 else 0 +
-            if (!task.categories.isNullOrBlank()) 72 else 0 +
-            if (task.completedAtMillis != null) 72 else 0 +
-            if (!task.syncError.isNullOrBlank()) 72 else 0 +
-            26
-        height.dp
-    }
-}
-
-private fun CollectionEntity.estimatedSettingsHeight(): Dp =
-    392.dp
-
-private fun detailTextBlockHeight(value: String, baseWhenBlank: Int = 72): Int {
-    val cleaned = value.cleanCalendarDisplayText()
-    if (cleaned.isBlank()) return baseWhenBlank
-    val explicitLines = cleaned.count { it == '\n' } + 1
-    val estimatedWrappedLines = (cleaned.length / 34) + 1
-    val lines = max(explicitLines, estimatedWrappedLines).coerceIn(1, 24)
-    return 54 + lines * 21
-}
-
-private fun EventEntity.estimatedParticipantsDetailHeight(): Int {
-    val organizer = organizerJson.toCalendarParticipant()
-    val attendees = attendeesJson.toCalendarParticipants()
-    if (organizer == null && attendees.isEmpty()) return 0
-
-    val header = 14
-    val surfacePadding = 22
-    val organizerHeight = if (organizer != null) 27 else 0
-    val attendeeRows = attendees.size * 44
-    val ownParticipationControls = if (attendees.isNotEmpty()) 78 else 0
-    val spacingCount = listOf(
-        1, // title to first content
-        if (organizer != null && attendees.isNotEmpty()) 1 else 0,
-        attendees.size.coerceAtLeast(0),
-        if (ownParticipationControls > 0) 2 else 0,
-    ).sum()
-    return surfacePadding + header + organizerHeight + attendeeRows + ownParticipationControls + spacingCount * 10
 }
 
 internal data class DraftEventSelection(
@@ -9673,11 +7088,6 @@ internal fun RecurrenceOption.localizedIntervalUnitLabel(): String = when (this)
     RecurrenceOption.Monthly -> appString(R.string.months)
     RecurrenceOption.Yearly -> appString(R.string.years)
     else -> appString(R.string.intervals)
-}
-
-sealed interface DetailSheet {
-    data class Event(val event: EventEntity) : DetailSheet
-    data class Task(val task: TaskEntity) : DetailSheet
 }
 
 internal data class EditorTransferDraft(
@@ -9829,7 +7239,7 @@ internal fun EventEntity.localizedAgendaSpanLabel(startDate: LocalDate, endDate:
 }
 
 @Composable
-private fun EventEntity.localizedFullTimeLabel(): String {
+internal fun EventEntity.localizedFullTimeLabel(): String {
     val dateFormatter = DateTimeFormatter.ofPattern("EEE, d. MMM yyyy", LocalAppLocale.current)
     val allDayLabel = stringResource(R.string.all_day)
     val startDate = startsAtMillis.toDate()
